Transform a gt table to a gtable object

as_gtable() performs the transformation of a gt_tbl object to a gtable object.

Usage

as_gtable(data, plot = FALSE, text_grob = grid::textGrob)

Arguments

data

The gt table data object

obj:<gt_tbl> // required

This is the gt table object that is commonly created through use of the gt() function.

plot

Render through the graphics device?

scalar<logical> // default: FALSE

The plot option determines whether the gtable object should be rendered on the graphics device.

text_grob

function // default: grid::textGrob

A function used to draw text. Defaults to grid::textGrob() but can be swapped to gridtext::richtext_grob() to better render HTML content.

Value

A gtable object.

Function ID

13-6

Function Introduced

In Development

See also