Output a gt object as RTF

Get the RTF content from a gt_tbl object as as a single-element character vector. This object can be used with writeLines() to generate a valid .rtf file that can be opened by RTF readers.

Usage

as_rtf(
  data,
  incl_open = TRUE,
  incl_header = TRUE,
  incl_page_info = TRUE,
  incl_body = TRUE,
  incl_close = TRUE
)

Arguments

data

The gt table data object

obj:<gt_tbl> // required

This is the gt table object that is commonly created through use of the gt() function.

incl_open, incl_close

Include opening/closing braces

scalar<logical> // default: TRUE

Options that govern whether the opening or closing "{" and "}" should be included. By default, both options are TRUE.

incl_header

Include RTF header

scalar<logical> // default: TRUE

Should the RTF header be included in the output? By default, this is TRUE.

incl_page_info

Include RTF page information

scalar<logical> // default: TRUE

Should the RTF output include directives for the document pages? This is TRUE by default.

incl_body

Include RTF body

scalar<logical> // default: TRUE

An option to include the body of RTF document. By default, this is TRUE.

Examples

Use a subset of the gtcars dataset to create a gt table. Add a header with tab_header() and then export the table as RTF code using the as_rtf() function.

tab_rtf <-
  gtcars |>
  dplyr::select(mfr, model) |>
  dplyr::slice(1:2) |>
  gt() |>
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) |>
  as_rtf()

Function ID

13-4

Function Introduced

v0.2.0.5 (March 31, 2020)

See also