Set columns to be aligned to the center

Set columns to be aligned to the center

cols_align_center(data, columns)

Arguments

data
a table object that is created using the gt() function.

columns
a vector of column names for which the alignment should be applied.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `mtcars` dataset and center align
# the `drat` column
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) %>%
    cols_align_center(columns = vars(drat))