Merge two columns to a single column with a formatter

Merge two columns to a single column with a formatter

cols_merge(data, col_1, col_2, pattern = "{1} {2}")

Arguments

data
a table object that is created using the gt() function.

col_1
a column that contains values for the start of the range.

col_2
a column that contains values for the end of the range.

pattern
a formatting pattern that specifies the arrangement of the col_1 and col_1 values and any string literals. The col_1 column is represented as {1} whereas col_2 is {2}. All other characters are taken to be string literals.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `sleep` dataset and merge the `group`
# and `ID` columns together (where the
# `ID` number is in parentheses after
# the `group` number
gt(sleep) %>%
  cols_merge(
    col_1 = vars(group),
    col_2 = vars(ID),
    pattern = "{1} ({2})")

#> <!--gt table start-->
#> <table data-nrow='20' data-ncol='2' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th class='col_heading center' rowspan='1' colspan='1'>extra</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>group</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='row center' style=''>0.7</td>
#> <td class='row center' style=''>1 (1)</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='row center' style=''>-1.6</td>
#> <td class='row center' style=''>1 (2)</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='row center' style=''>-0.2</td>
#> <td class='row center' style=''>1 (3)</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='row center' style=''>-1.2</td>
#> <td class='row center' style=''>1 (4)</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='row center' style=''>-0.1</td>
#> <td class='row center' style=''>1 (5)</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1 (6)</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='row center' style=''>3.7</td>
#> <td class='row center' style=''>1 (7)</td>
#> </tr>
#> <tr data-type='data' data-row='8'>
#> <td class='row center' style=''>0.8</td>
#> <td class='row center' style=''>1 (8)</td>
#> </tr>
#> <tr data-type='data' data-row='9'>
#> <td class='row center' style=''>0</td>
#> <td class='row center' style=''>1 (9)</td>
#> </tr>
#> <tr data-type='data' data-row='10'>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>1 (10)</td>
#> </tr>
#> <tr data-type='data' data-row='11'>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>2 (1)</td>
#> </tr>
#> <tr data-type='data' data-row='12'>
#> <td class='row center' style=''>0.8</td>
#> <td class='row center' style=''>2 (2)</td>
#> </tr>
#> <tr data-type='data' data-row='13'>
#> <td class='row center' style=''>1.1</td>
#> <td class='row center' style=''>2 (3)</td>
#> </tr>
#> <tr data-type='data' data-row='14'>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>2 (4)</td>
#> </tr>
#> <tr data-type='data' data-row='15'>
#> <td class='row center' style=''>-0.1</td>
#> <td class='row center' style=''>2 (5)</td>
#> </tr>
#> <tr data-type='data' data-row='16'>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>2 (6)</td>
#> </tr>
#> <tr data-type='data' data-row='17'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2 (7)</td>
#> </tr>
#> <tr data-type='data' data-row='18'>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>2 (8)</td>
#> </tr>
#> <tr data-type='data' data-row='19'>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>2 (9)</td>
#> </tr>
#> <tr data-type='data' data-row='20'>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>2 (10)</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>