Create group names and column labels via delimited column names

This option will split delimited column names such that the first component is promoted to the group name (and hence will span over column labels) and subsequent components will represent the column label. Please note that reference to individual columns must still be via the original column names.

cols_split_delim(data, delim, columns = NULL)

Arguments

data
a table object that is created using the gt() function.

delim
the delimiter to use to split an input column name. The delimiter supplied will be autoescaped for the internal splitting procedure. The first component of the split will become the group name and the second component will be the column label.

columns
an optional vector of column names that this operation should be limited to. The default is to consider all columns in the table.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `iris` dataset and split any columns
# that are dot separated between column
# groups and column labels.
gt(iris) %>%
  cols_split_delim(delim = ".")

#> <!--gt table start-->
#> <table data-nrow='150' data-ncol='5' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th data-type='column_heading' class='col_heading column_spanner sep_right center' rowspan='1' colspan='2'>Sepal</th>
#> <th data-type='column_heading' class='col_heading column_spanner center' rowspan='1' colspan='2'>Petal</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='2' colspan='1'>Species</th>
#> </tr>
#> <tr>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>Length</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>Width</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>Length</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>Width</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.6</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3.9</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='8'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='9'>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='10'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='11'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3.7</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='12'>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='13'>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='14'>
#> <td class='row center' style=''>4.3</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.1</td>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='15'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='16'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='17'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3.9</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='18'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='19'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='20'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='21'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='22'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.7</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='23'>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>3.6</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='24'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>0.5</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='25'>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='26'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='27'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='28'>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='29'>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='30'>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='31'>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='32'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='33'>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>4.1</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='34'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>4.2</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='35'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='36'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='37'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='38'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>3.6</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.1</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='39'>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='40'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='41'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='42'>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='43'>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='44'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.6</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='45'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>0.4</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='46'>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.3</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='47'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='48'>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='49'>
#> <td class='row center' style=''>5.3</td>
#> <td class='row center' style=''>3.7</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='50'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>0.2</td>
#> <td class='row center' style=''>setosa</td>
#> </tr>
#> <tr data-type='data' data-row='51'>
#> <td class='row center' style=''>7</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='52'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='53'>
#> <td class='row center' style=''>6.9</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='54'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='55'>
#> <td class='row center' style=''>6.5</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='56'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='57'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='58'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='59'>
#> <td class='row center' style=''>6.6</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='60'>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>3.9</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='61'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='62'>
#> <td class='row center' style=''>5.9</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.2</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='63'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='64'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='65'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>3.6</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='66'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='67'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='68'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>4.1</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='69'>
#> <td class='row center' style=''>6.2</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='70'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>3.9</td>
#> <td class='row center' style=''>1.1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='71'>
#> <td class='row center' style=''>5.9</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='72'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='73'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='74'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='75'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.3</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='76'>
#> <td class='row center' style=''>6.6</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='77'>
#> <td class='row center' style=''>6.8</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='78'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='79'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='80'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.6</td>
#> <td class='row center' style=''>3.5</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='81'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>1.1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='82'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>3.7</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='83'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>3.9</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='84'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='85'>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='86'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='87'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>4.7</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='88'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='89'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.1</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='90'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='91'>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.6</td>
#> <td class='row center' style=''>4.4</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='92'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.6</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='93'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.6</td>
#> <td class='row center' style=''>4</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='94'>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='95'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>4.2</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='96'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.2</td>
#> <td class='row center' style=''>1.2</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='97'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.2</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='98'>
#> <td class='row center' style=''>6.2</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>4.3</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='99'>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>1.1</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='100'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.1</td>
#> <td class='row center' style=''>1.3</td>
#> <td class='row center' style=''>versicolor</td>
#> </tr>
#> <tr data-type='data' data-row='101'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='102'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='103'>
#> <td class='row center' style=''>7.1</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.9</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='104'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='105'>
#> <td class='row center' style=''>6.5</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='106'>
#> <td class='row center' style=''>7.6</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>6.6</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='107'>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>4.5</td>
#> <td class='row center' style=''>1.7</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='108'>
#> <td class='row center' style=''>7.3</td>
#> <td class='row center' style=''>2.9</td>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='109'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='110'>
#> <td class='row center' style=''>7.2</td>
#> <td class='row center' style=''>3.6</td>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='111'>
#> <td class='row center' style=''>6.5</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='112'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>5.3</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='113'>
#> <td class='row center' style=''>6.8</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='114'>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='115'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='116'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>5.3</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='117'>
#> <td class='row center' style=''>6.5</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='118'>
#> <td class='row center' style=''>7.7</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='119'>
#> <td class='row center' style=''>7.7</td>
#> <td class='row center' style=''>2.6</td>
#> <td class='row center' style=''>6.9</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='120'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='121'>
#> <td class='row center' style=''>6.9</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='122'>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='123'>
#> <td class='row center' style=''>7.7</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='124'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='125'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='126'>
#> <td class='row center' style=''>7.2</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='127'>
#> <td class='row center' style=''>6.2</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='128'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.9</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='129'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='130'>
#> <td class='row center' style=''>7.2</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>1.6</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='131'>
#> <td class='row center' style=''>7.4</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='132'>
#> <td class='row center' style=''>7.9</td>
#> <td class='row center' style=''>3.8</td>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='133'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='134'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.8</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>1.5</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='135'>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.6</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>1.4</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='136'>
#> <td class='row center' style=''>7.7</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>6.1</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='137'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='138'>
#> <td class='row center' style=''>6.4</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>5.5</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='139'>
#> <td class='row center' style=''>6</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>4.8</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='140'>
#> <td class='row center' style=''>6.9</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>2.1</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='141'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>5.6</td>
#> <td class='row center' style=''>2.4</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='142'>
#> <td class='row center' style=''>6.9</td>
#> <td class='row center' style=''>3.1</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='143'>
#> <td class='row center' style=''>5.8</td>
#> <td class='row center' style=''>2.7</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='144'>
#> <td class='row center' style=''>6.8</td>
#> <td class='row center' style=''>3.2</td>
#> <td class='row center' style=''>5.9</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='145'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3.3</td>
#> <td class='row center' style=''>5.7</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='146'>
#> <td class='row center' style=''>6.7</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='147'>
#> <td class='row center' style=''>6.3</td>
#> <td class='row center' style=''>2.5</td>
#> <td class='row center' style=''>5</td>
#> <td class='row center' style=''>1.9</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='148'>
#> <td class='row center' style=''>6.5</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.2</td>
#> <td class='row center' style=''>2</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='149'>
#> <td class='row center' style=''>6.2</td>
#> <td class='row center' style=''>3.4</td>
#> <td class='row center' style=''>5.4</td>
#> <td class='row center' style=''>2.3</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> <tr data-type='data' data-row='150'>
#> <td class='row center' style=''>5.9</td>
#> <td class='row center' style=''>3</td>
#> <td class='row center' style=''>5.1</td>
#> <td class='row center' style=''>1.8</td>
#> <td class='row center' style=''>virginica</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>