escape_latex {gt}	R Documentation

Perform LaTeX escaping

Description

Text may contain several characters with special meanings in LaTeX. This function will transform a character vector so that it is safe to use within LaTeX tables.

Usage

escape_latex(text)

Arguments

text
a character vector containing the text that is to be LaTeX-escaped.

See Also

Other helper functions: cells_styles, gt_latex_dependencies, html, md, pct, px

[Package gt version 0.1.0 Index]