Format values as dates

Format input date values that are character-based and expressed according to the ISO 8601 date format (YYYY-MM-DD). Once the appropriate data cells are targeted with columns (and, optionally, rows), we can simply apply a preset date style to format the dates.

fmt_date(data, columns, rows = NULL, date_style)

Arguments

data
a table object that is created using the gt() function.

columns
the columns to format. Can either be a series of column names provided in vars(), a vector of column indices, or a helper function focused on selections. The select helper functions are: starts_with(), ends_with(), contains(), matches(), one_of(), and everything().

rows
optional rows to format. Not providing any value results in all rows in columns being formatted. Can either be a vector of row captions provided c(), a vector of row indices, or a helper function focused on selections. The select helper functions are: starts_with(), ends_with(), contains(), matches(), one_of(), and everything(). We can also use expressions to filter down to the rows we need (e.g., [colname_1] > 100 & [colname_2] < 50).

date_style
the date style to use. Supply a number (from 1 to 14) that corresponds to the preferred date style. Use info_date_style() to see the different numbered and named date presets.

Value

an object of class gt_tbl.

Details

The following date styles are available for simpler formatting of ISO dates (all using the input date of 2000-02-29 in the example output dates):

iso: 2000-02-29

wday_month_day_year: Tuesday, February 29, 2000

wd_m_day_year: Tue, Feb 29, 2000

wday_day_month_year: Tuesday 29 February 2000

month_day_year: February 29, 2000

m_day_year: Feb 29, 2000

day_m_year: 29 Feb 2000

day_month_year: 29 February 2000

day_month: 29 February

year: 2000

month: February

day: 29

year.mn.day: 2000/02/29

y.mn.day: 0/02/29

We can use the info_date_style() function for a useful reference on all of the possible inputs to date_style.

See also

Examples

# Create a tibble with a column
# that contains dates
data_tbl <-
  dplyr::tribble(
    ~date,
    "2017-10-15",
    "2013-02-22",
    "2014-09-22",
    "2018-01-10")

# Create a table object using this
# dataset and format the `date`
# column with a `date_style` of `2`
gt_tbl <-
  gt(data_tbl) %>%
    fmt_date(
      columns = vars(date),
      date_style = 2)