Format numeric values

Format numeric values

fmt_number(data, columns, rows = NULL, decimals = 2,
  drop_trailing_zeros = FALSE, negative_val = "signed",
  use_seps = TRUE, scale_by = 1, pattern = "{x}", sep_mark = ",",
  dec_mark = ".", locale = NULL)

Arguments

data
a table object that is created using the gt() function.

columns
the column names to format.

rows
optional rows to format. Not providing any value results in all rows in columns being formatted.

decimals
an option to specify the exact number of decimal places to use. The default number of decimal places is 2.

drop_trailing_zeros
a logical value that allows for removal of trailing zeros (those redundant zeros after the decimal mark).

negative_val
the formatting to use for negative values. With signed (the default), negative values will be shown with a negative sign. Using parens will show the negative value in parentheses.

use_seps
an option to use digit group separators. The type of digit group separator is set by sep_mark and overridden if a locale ID is provided to locale. This setting is TRUE by default.

scale_by
a value to scale the input. The default is 1.0.

pattern
a formatting pattern that allows for decoration of the formatted value. The value itself is represented by {x} and all other characters are taken to be string literals.

sep_mark
the mark to use as a separator between groups of digits.

dec_mark
the character to use as a decimal mark.

locale
an optional locale ID that can be used for formatting the value according the locale's rules. Examples include "en_US" for English (United States) and "fr_FR" for French (France). The use of a valid locale ID will override any values provided in sep_mark and dec_mark.

Value

an object of class gt_tbl.

See also

Examples

library(tidyverse)

# Create a table object using the
# `mtcars` dataset and format specified
# numeric columns to display values to
# two decimal places
gt(mtcars, rownames_to_stub = TRUE) %>%
  fmt_number(
    columns = vars(mpg, disp, drat, qsec),
    decimals = 5)

#> <!--gt table start-->
#> <table data-nrow='32' data-ncol='12' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th class='col_heading right' rowspan='1' colspan='1'></th>
#> <th class='col_heading center' rowspan='1' colspan='1'>mpg</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>cyl</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>disp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>hp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>drat</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>wt</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>qsec</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>vs</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>am</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>gear</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>carb</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='stub row right'>Mazda RX4</td>
#> <td class='row center'>21.00000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>160.00000</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.90000</td>
#> <td class='row center'>2.62</td>
#> <td class='row center'>16.46000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='stub row right'>Mazda RX4 Wag</td>
#> <td class='row center'>21.00000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>160.00000</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.90000</td>
#> <td class='row center'>2.875</td>
#> <td class='row center'>17.02000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='stub row right'>Datsun 710</td>
#> <td class='row center'>22.80000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>108.00000</td>
#> <td class='row center'>93</td>
#> <td class='row center'>3.85000</td>
#> <td class='row center'>2.32</td>
#> <td class='row center'>18.61000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='stub row right'>Hornet 4 Drive</td>
#> <td class='row center'>21.40000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>258.00000</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.08000</td>
#> <td class='row center'>3.215</td>
#> <td class='row center'>19.44000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='stub row right'>Hornet Sportabout</td>
#> <td class='row center'>18.70000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>360.00000</td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.15000</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>17.02000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='stub row right'>Valiant</td>
#> <td class='row center'>18.10000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>225.00000</td>
#> <td class='row center'>105</td>
#> <td class='row center'>2.76000</td>
#> <td class='row center'>3.46</td>
#> <td class='row center'>20.22000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='stub row right'>Duster 360</td>
#> <td class='row center'>14.30000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>360.00000</td>
#> <td class='row center'>245</td>
#> <td class='row center'>3.21000</td>
#> <td class='row center'>3.57</td>
#> <td class='row center'>15.84000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='8'>
#> <td class='stub row right'>Merc 240D</td>
#> <td class='row center'>24.40000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>146.70000</td>
#> <td class='row center'>62</td>
#> <td class='row center'>3.69000</td>
#> <td class='row center'>3.19</td>
#> <td class='row center'>20.00000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='9'>
#> <td class='stub row right'>Merc 230</td>
#> <td class='row center'>22.80000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>140.80000</td>
#> <td class='row center'>95</td>
#> <td class='row center'>3.92000</td>
#> <td class='row center'>3.15</td>
#> <td class='row center'>22.90000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='10'>
#> <td class='stub row right'>Merc 280</td>
#> <td class='row center'>19.20000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>167.60000</td>
#> <td class='row center'>123</td>
#> <td class='row center'>3.92000</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>18.30000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='11'>
#> <td class='stub row right'>Merc 280C</td>
#> <td class='row center'>17.80000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>167.60000</td>
#> <td class='row center'>123</td>
#> <td class='row center'>3.92000</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>18.90000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='12'>
#> <td class='stub row right'>Merc 450SE</td>
#> <td class='row center'>16.40000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>275.80000</td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07000</td>
#> <td class='row center'>4.07</td>
#> <td class='row center'>17.40000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='13'>
#> <td class='stub row right'>Merc 450SL</td>
#> <td class='row center'>17.30000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>275.80000</td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07000</td>
#> <td class='row center'>3.73</td>
#> <td class='row center'>17.60000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='14'>
#> <td class='stub row right'>Merc 450SLC</td>
#> <td class='row center'>15.20000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>275.80000</td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07000</td>
#> <td class='row center'>3.78</td>
#> <td class='row center'>18.00000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='15'>
#> <td class='stub row right'>Cadillac Fleetwood</td>
#> <td class='row center'>10.40000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>472.00000</td>
#> <td class='row center'>205</td>
#> <td class='row center'>2.93000</td>
#> <td class='row center'>5.25</td>
#> <td class='row center'>17.98000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='16'>
#> <td class='stub row right'>Lincoln Continental</td>
#> <td class='row center'>10.40000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>460.00000</td>
#> <td class='row center'>215</td>
#> <td class='row center'>3.00000</td>
#> <td class='row center'>5.424</td>
#> <td class='row center'>17.82000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='17'>
#> <td class='stub row right'>Chrysler Imperial</td>
#> <td class='row center'>14.70000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>440.00000</td>
#> <td class='row center'>230</td>
#> <td class='row center'>3.23000</td>
#> <td class='row center'>5.345</td>
#> <td class='row center'>17.42000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='18'>
#> <td class='stub row right'>Fiat 128</td>
#> <td class='row center'>32.40000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>78.70000</td>
#> <td class='row center'>66</td>
#> <td class='row center'>4.08000</td>
#> <td class='row center'>2.2</td>
#> <td class='row center'>19.47000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='19'>
#> <td class='stub row right'>Honda Civic</td>
#> <td class='row center'>30.40000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>75.70000</td>
#> <td class='row center'>52</td>
#> <td class='row center'>4.93000</td>
#> <td class='row center'>1.615</td>
#> <td class='row center'>18.52000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='20'>
#> <td class='stub row right'>Toyota Corolla</td>
#> <td class='row center'>33.90000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>71.10000</td>
#> <td class='row center'>65</td>
#> <td class='row center'>4.22000</td>
#> <td class='row center'>1.835</td>
#> <td class='row center'>19.90000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='21'>
#> <td class='stub row right'>Toyota Corona</td>
#> <td class='row center'>21.50000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>120.10000</td>
#> <td class='row center'>97</td>
#> <td class='row center'>3.70000</td>
#> <td class='row center'>2.465</td>
#> <td class='row center'>20.01000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='22'>
#> <td class='stub row right'>Dodge Challenger</td>
#> <td class='row center'>15.50000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>318.00000</td>
#> <td class='row center'>150</td>
#> <td class='row center'>2.76000</td>
#> <td class='row center'>3.52</td>
#> <td class='row center'>16.87000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='23'>
#> <td class='stub row right'>AMC Javelin</td>
#> <td class='row center'>15.20000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>304.00000</td>
#> <td class='row center'>150</td>
#> <td class='row center'>3.15000</td>
#> <td class='row center'>3.435</td>
#> <td class='row center'>17.30000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='24'>
#> <td class='stub row right'>Camaro Z28</td>
#> <td class='row center'>13.30000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>350.00000</td>
#> <td class='row center'>245</td>
#> <td class='row center'>3.73000</td>
#> <td class='row center'>3.84</td>
#> <td class='row center'>15.41000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='25'>
#> <td class='stub row right'>Pontiac Firebird</td>
#> <td class='row center'>19.20000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>400.00000</td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.08000</td>
#> <td class='row center'>3.845</td>
#> <td class='row center'>17.05000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='26'>
#> <td class='stub row right'>Fiat X1-9</td>
#> <td class='row center'>27.30000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>79.00000</td>
#> <td class='row center'>66</td>
#> <td class='row center'>4.08000</td>
#> <td class='row center'>1.935</td>
#> <td class='row center'>18.90000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='27'>
#> <td class='stub row right'>Porsche 914-2</td>
#> <td class='row center'>26.00000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>120.30000</td>
#> <td class='row center'>91</td>
#> <td class='row center'>4.43000</td>
#> <td class='row center'>2.14</td>
#> <td class='row center'>16.70000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='28'>
#> <td class='stub row right'>Lotus Europa</td>
#> <td class='row center'>30.40000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>95.10000</td>
#> <td class='row center'>113</td>
#> <td class='row center'>3.77000</td>
#> <td class='row center'>1.513</td>
#> <td class='row center'>16.90000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='29'>
#> <td class='stub row right'>Ford Pantera L</td>
#> <td class='row center'>15.80000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>351.00000</td>
#> <td class='row center'>264</td>
#> <td class='row center'>4.22000</td>
#> <td class='row center'>3.17</td>
#> <td class='row center'>14.50000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='30'>
#> <td class='stub row right'>Ferrari Dino</td>
#> <td class='row center'>19.70000</td>
#> <td class='row center'>6</td>
#> <td class='row center'>145.00000</td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.62000</td>
#> <td class='row center'>2.77</td>
#> <td class='row center'>15.50000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>6</td>
#> </tr>
#> <tr data-type='data' data-row='31'>
#> <td class='stub row right'>Maserati Bora</td>
#> <td class='row center'>15.00000</td>
#> <td class='row center'>8</td>
#> <td class='row center'>301.00000</td>
#> <td class='row center'>335</td>
#> <td class='row center'>3.54000</td>
#> <td class='row center'>3.57</td>
#> <td class='row center'>14.60000</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>8</td>
#> </tr>
#> <tr data-type='data' data-row='32'>
#> <td class='stub row right'>Volvo 142E</td>
#> <td class='row center'>21.40000</td>
#> <td class='row center'>4</td>
#> <td class='row center'>121.00000</td>
#> <td class='row center'>109</td>
#> <td class='row center'>4.11000</td>
#> <td class='row center'>2.78</td>
#> <td class='row center'>18.60000</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#> 

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_1,   ~val_2,
    8236625,  -2345,
    6533405,  -7865,
    7363931,  2345,
    9326440,  7543)

# Create a gt table from `data_tbl` and
# scale the numeric values in columns
# `val_1` and `val_2` and apply a suffix
gt(data_tbl) %>%
  fmt_number(
    columns = vars(val_1),
    decimals = 2,
    scale_by = 1E-6,
    pattern = "{x}") %>%
  fmt_number(
    columns = vars(val_2),
    decimals = 2,
    scale_by = 1/1000,
    pattern = "{x}K")

#> <!--gt table start-->
#> <table data-nrow='4' data-ncol='2' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th class='col_heading center' rowspan='1' colspan='1'>val_1</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>val_2</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='row center' style=''>8.24</td>
#> <td class='row center' style=''>-2.35K</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='row center' style=''>6.53</td>
#> <td class='row center' style=''>-7.87K</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='row center' style=''>7.36</td>
#> <td class='row center' style=''>2.35K</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='row center' style=''>9.33</td>
#> <td class='row center' style=''>7.54K</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>