Format values to scientific notation

Format values to scientific notation

fmt_scientific(data, columns, rows = NULL, decimals = 2,
  drop_trailing_zeros = FALSE, scale_by = 1, pattern = "{x}",
  sep_mark = ",", dec_mark = ".", locale = NULL)

Arguments

data
a table object that is created using the gt() function.

columns
the column names to format.

rows
optional rows to format. Not providing any value results in all rows in columns being formatted.

decimals
an option to specify the exact number of decimal places to use. The default number of decimal places is 2.

drop_trailing_zeros
a logical value that allows for removal of trailing zeros (those redundant zeros after the decimal mark).

scale_by
a value to scale the input. The default is 1.0.

pattern
a formatting pattern that allows for decoration of the formatted value. The value itself is represented by {x} and all other characters are taken to be string literals.

sep_mark
the mark to use as a separator between groups of digits.

dec_mark
the character to use as a decimal mark.

locale
an optional locale ID that can be used for formatting the value according the locale's rules. Examples include "en_US" for English (United States) and "fr_FR" for French (France). The use of a valid locale ID will override any values provided in sep_mark and dec_mark.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `mtcars` dataset and format specified
# numeric columns to display values in
# scientific notation
gt(mtcars, rownames_to_stub = TRUE) %>%
  fmt_scientific(
    columns = vars(disp),
    decimals = 2)

#> <style>/* gt compiled CSS file */
#> .gt_table {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#>   border-collapse: collapse;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: 100%;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> .heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#> }
#> 
#> .title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#> }
#> 
#> .headnote {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.headnote.font.size */
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#> }
#> 
#> .bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> .column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> .col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* boxhead.background.color */
#>   font-size: 16px;
#>   /* boxhead.font.size */
#>   font-weight: initial;
#>   /* boxhead.font.weight */
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> /* this is used to create a gutter between adjacent column spanners;
#>    the color should match that of .heading's background-color */
#> .sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> .group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#> }
#> 
#> .striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> .row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#> }
#> 
#> .stub {
#>   width: 10%;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#> }
#> 
#> .stub.row.right, .stub.row.left, .stub.row.center {
#>   background-color: #FFFFFF;
#> }
#> 
#> .summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> .first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> .table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> .footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> .sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> .center {
#>   text-align: center;
#> }
#> 
#> .left {
#>   text-align: left;
#> }
#> 
#> .right {
#>   text-align: right;
#> }
#> 
#> .font_normal {
#>   font-weight: normal;
#> }
#> 
#> .font_bold {
#>   font-weight: bold;
#> }
#> 
#> .font_italic {
#>   font-style: italic;
#> }
#> 
#> .gt_super {
#>   font-size: 65%;
#> }
#> </style>
#> <!--gt table start-->
#> <table data-nrow='32' data-ncol='12' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th class='col_heading right' rowspan='1' colspan='1'></th>
#> <th class='col_heading center' rowspan='1' colspan='1'>mpg</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>cyl</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>disp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>hp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>drat</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>wt</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>qsec</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>vs</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>am</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>gear</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>carb</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='stub row right'>Mazda RX4</td>
#> <td class='row center'>21</td>
#> <td class='row center'>6</td>
#> <td class='row center'>1.60 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.9</td>
#> <td class='row center'>2.62</td>
#> <td class='row center'>16.46</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='stub row right'>Mazda RX4 Wag</td>
#> <td class='row center'>21</td>
#> <td class='row center'>6</td>
#> <td class='row center'>1.60 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.9</td>
#> <td class='row center'>2.875</td>
#> <td class='row center'>17.02</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='stub row right'>Datsun 710</td>
#> <td class='row center'>22.8</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.08 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>93</td>
#> <td class='row center'>3.85</td>
#> <td class='row center'>2.32</td>
#> <td class='row center'>18.61</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='stub row right'>Hornet 4 Drive</td>
#> <td class='row center'>21.4</td>
#> <td class='row center'>6</td>
#> <td class='row center'>2.58 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.08</td>
#> <td class='row center'>3.215</td>
#> <td class='row center'>19.44</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='stub row right'>Hornet Sportabout</td>
#> <td class='row center'>18.7</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.60 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.15</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>17.02</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='stub row right'>Valiant</td>
#> <td class='row center'>18.1</td>
#> <td class='row center'>6</td>
#> <td class='row center'>2.25 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>105</td>
#> <td class='row center'>2.76</td>
#> <td class='row center'>3.46</td>
#> <td class='row center'>20.22</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='stub row right'>Duster 360</td>
#> <td class='row center'>14.3</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.60 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>245</td>
#> <td class='row center'>3.21</td>
#> <td class='row center'>3.57</td>
#> <td class='row center'>15.84</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='8'>
#> <td class='stub row right'>Merc 240D</td>
#> <td class='row center'>24.4</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.47 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>62</td>
#> <td class='row center'>3.69</td>
#> <td class='row center'>3.19</td>
#> <td class='row center'>20</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='9'>
#> <td class='stub row right'>Merc 230</td>
#> <td class='row center'>22.8</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.41 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>95</td>
#> <td class='row center'>3.92</td>
#> <td class='row center'>3.15</td>
#> <td class='row center'>22.9</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='10'>
#> <td class='stub row right'>Merc 280</td>
#> <td class='row center'>19.2</td>
#> <td class='row center'>6</td>
#> <td class='row center'>1.68 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>123</td>
#> <td class='row center'>3.92</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>18.3</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='11'>
#> <td class='stub row right'>Merc 280C</td>
#> <td class='row center'>17.8</td>
#> <td class='row center'>6</td>
#> <td class='row center'>1.68 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>123</td>
#> <td class='row center'>3.92</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>18.9</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='12'>
#> <td class='stub row right'>Merc 450SE</td>
#> <td class='row center'>16.4</td>
#> <td class='row center'>8</td>
#> <td class='row center'>2.76 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07</td>
#> <td class='row center'>4.07</td>
#> <td class='row center'>17.4</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='13'>
#> <td class='stub row right'>Merc 450SL</td>
#> <td class='row center'>17.3</td>
#> <td class='row center'>8</td>
#> <td class='row center'>2.76 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07</td>
#> <td class='row center'>3.73</td>
#> <td class='row center'>17.6</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='14'>
#> <td class='stub row right'>Merc 450SLC</td>
#> <td class='row center'>15.2</td>
#> <td class='row center'>8</td>
#> <td class='row center'>2.76 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>180</td>
#> <td class='row center'>3.07</td>
#> <td class='row center'>3.78</td>
#> <td class='row center'>18</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>3</td>
#> </tr>
#> <tr data-type='data' data-row='15'>
#> <td class='stub row right'>Cadillac Fleetwood</td>
#> <td class='row center'>10.4</td>
#> <td class='row center'>8</td>
#> <td class='row center'>4.72 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>205</td>
#> <td class='row center'>2.93</td>
#> <td class='row center'>5.25</td>
#> <td class='row center'>17.98</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='16'>
#> <td class='stub row right'>Lincoln Continental</td>
#> <td class='row center'>10.4</td>
#> <td class='row center'>8</td>
#> <td class='row center'>4.60 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>215</td>
#> <td class='row center'>3</td>
#> <td class='row center'>5.424</td>
#> <td class='row center'>17.82</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='17'>
#> <td class='stub row right'>Chrysler Imperial</td>
#> <td class='row center'>14.7</td>
#> <td class='row center'>8</td>
#> <td class='row center'>4.40 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>230</td>
#> <td class='row center'>3.23</td>
#> <td class='row center'>5.345</td>
#> <td class='row center'>17.42</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='18'>
#> <td class='stub row right'>Fiat 128</td>
#> <td class='row center'>32.4</td>
#> <td class='row center'>4</td>
#> <td class='row center'>7.87 &times; 10<sup class='gt_super'>1</sup></td>
#> <td class='row center'>66</td>
#> <td class='row center'>4.08</td>
#> <td class='row center'>2.2</td>
#> <td class='row center'>19.47</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='19'>
#> <td class='stub row right'>Honda Civic</td>
#> <td class='row center'>30.4</td>
#> <td class='row center'>4</td>
#> <td class='row center'>7.57 &times; 10<sup class='gt_super'>1</sup></td>
#> <td class='row center'>52</td>
#> <td class='row center'>4.93</td>
#> <td class='row center'>1.615</td>
#> <td class='row center'>18.52</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='20'>
#> <td class='stub row right'>Toyota Corolla</td>
#> <td class='row center'>33.9</td>
#> <td class='row center'>4</td>
#> <td class='row center'>7.11 &times; 10<sup class='gt_super'>1</sup></td>
#> <td class='row center'>65</td>
#> <td class='row center'>4.22</td>
#> <td class='row center'>1.835</td>
#> <td class='row center'>19.9</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='21'>
#> <td class='stub row right'>Toyota Corona</td>
#> <td class='row center'>21.5</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.20 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>97</td>
#> <td class='row center'>3.7</td>
#> <td class='row center'>2.465</td>
#> <td class='row center'>20.01</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='22'>
#> <td class='stub row right'>Dodge Challenger</td>
#> <td class='row center'>15.5</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.18 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>150</td>
#> <td class='row center'>2.76</td>
#> <td class='row center'>3.52</td>
#> <td class='row center'>16.87</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='23'>
#> <td class='stub row right'>AMC Javelin</td>
#> <td class='row center'>15.2</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.04 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>150</td>
#> <td class='row center'>3.15</td>
#> <td class='row center'>3.435</td>
#> <td class='row center'>17.3</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='24'>
#> <td class='stub row right'>Camaro Z28</td>
#> <td class='row center'>13.3</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.50 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>245</td>
#> <td class='row center'>3.73</td>
#> <td class='row center'>3.84</td>
#> <td class='row center'>15.41</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='25'>
#> <td class='stub row right'>Pontiac Firebird</td>
#> <td class='row center'>19.2</td>
#> <td class='row center'>8</td>
#> <td class='row center'>4.00 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.08</td>
#> <td class='row center'>3.845</td>
#> <td class='row center'>17.05</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='26'>
#> <td class='stub row right'>Fiat X1-9</td>
#> <td class='row center'>27.3</td>
#> <td class='row center'>4</td>
#> <td class='row center'>7.90 &times; 10<sup class='gt_super'>1</sup></td>
#> <td class='row center'>66</td>
#> <td class='row center'>4.08</td>
#> <td class='row center'>1.935</td>
#> <td class='row center'>18.9</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='27'>
#> <td class='stub row right'>Porsche 914-2</td>
#> <td class='row center'>26</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.20 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>91</td>
#> <td class='row center'>4.43</td>
#> <td class='row center'>2.14</td>
#> <td class='row center'>16.7</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='28'>
#> <td class='stub row right'>Lotus Europa</td>
#> <td class='row center'>30.4</td>
#> <td class='row center'>4</td>
#> <td class='row center'>9.51 &times; 10<sup class='gt_super'>1</sup></td>
#> <td class='row center'>113</td>
#> <td class='row center'>3.77</td>
#> <td class='row center'>1.513</td>
#> <td class='row center'>16.9</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='29'>
#> <td class='stub row right'>Ford Pantera L</td>
#> <td class='row center'>15.8</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.51 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>264</td>
#> <td class='row center'>4.22</td>
#> <td class='row center'>3.17</td>
#> <td class='row center'>14.5</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='30'>
#> <td class='stub row right'>Ferrari Dino</td>
#> <td class='row center'>19.7</td>
#> <td class='row center'>6</td>
#> <td class='row center'>1.45 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.62</td>
#> <td class='row center'>2.77</td>
#> <td class='row center'>15.5</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>6</td>
#> </tr>
#> <tr data-type='data' data-row='31'>
#> <td class='stub row right'>Maserati Bora</td>
#> <td class='row center'>15</td>
#> <td class='row center'>8</td>
#> <td class='row center'>3.01 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>335</td>
#> <td class='row center'>3.54</td>
#> <td class='row center'>3.57</td>
#> <td class='row center'>14.6</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>8</td>
#> </tr>
#> <tr data-type='data' data-row='32'>
#> <td class='stub row right'>Volvo 142E</td>
#> <td class='row center'>21.4</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1.21 &times; 10<sup class='gt_super'>2</sup></td>
#> <td class='row center'>109</td>
#> <td class='row center'>4.11</td>
#> <td class='row center'>2.78</td>
#> <td class='row center'>18.6</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>2</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>