Reference a column of values for certain parameters

It can be useful to obtain parameter values from a column in a gt for functions that operate on the table body and stub cells. For example, you might want to indent row labels in the stub. You could call tab_stub_indent() and indent different rows to various indentation levels. However, each level of indentation applied necessitates a new call of that function. To make this better, we can use indentation values available in a table column via the from_column() helper function. For the tab_stub_indent() case, you'd invoke this helper at the indent argument and specify the column that has the values.

Usage

from_column(column, na_value = NULL, fn = NULL)

Arguments

column

Column name

scalar<character> // required

A single column name in quotation marks. Values will be extracted from this column and provided to compatible arguments.

na_value

Default replacement for NA values

scalar<character|numeric|logical> // default: NULL (optional)

A single value to replace any NA values in the column. Take care to provide a value that is of the same type as the column values to avoid any undesirable coercion.

fn

Function to apply

function|formula // default: NULL (optional)

If a function is provided here, any values extracted from the table column (except NA values) can be mutated.

Value

A list object of class gt_column.

Functions that allow the use of the from_column() helper

Only certain functions (and furthermore a subset of arguments within each) support the use of from_column() for accessing varying parameter values. These functions are:

tab_stub_indent()

fmt_number()

fmt_integer()

fmt_scientific()

fmt_engineering()

fmt_percent()

fmt_partsper()

fmt_fraction()

fmt_currency()

fmt_roman()

fmt_index()

fmt_spelled_num()

fmt_bytes()

fmt_date()

fmt_time()

fmt_datetime()

fmt_url()

fmt_image()

fmt_flag()

fmt_markdown()

fmt_passthrough()

Within help documents for each of these functions you'll find the Compatibility of arguments with the from_column() helper function section and sections like these describe which arguments support the use of from_column().

Function ID

8-5

Function Introduced

In Development

See also