gt-package {gt}	R Documentation

gt: Easily Create Presentation-Ready Display Tables

Description

Build display tables from tabular data using with an easy-to-use API. With its progressive approach, we can construct display tables with a clear separation of concerns: you don't have to decide how the tabular data gets transformed and structured whilst also worrying about aesthetics.

Author(s)

Maintainer: Richard Iannone rich@rstudio.com (0000-0003-3925-190X)

Authors:

Joe Cheng joe@rstudio.com

Barret Schloerke barret@rstudio.com (0000-0001-9986-114X)

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://github.com/rstudio/gt

Report bugs at https://github.com/rstudio/gt/issues

[Package gt version 0.1.0 Index]