An ADaM-flavored clinical trial toy dataset

This tibble contains artificial data for 182 subjects of the GT01 study. Each row corresponds to demographic characteristics of a single trial participant. Two out of 182 study participants were screen failures and thus not treated, the rest of the study population was randomized with a 1:1 ratio to receive either Placebo (as comparator) or Drug 1. The dataset entails subject level demographics such as age, age group, sex, ethnicity and body mass index (BMI) at baseline, as well as an event flag, indicating whether the subject experienced a specific event throughout the course of the study or not. Although the data was intentionally created to mimic a typical clinical trial dataset following the CDISC format, it might not strictly comply with CDISC ADaM rules. The intent is to showcase the workflow for clinical table creation rather than creating a fully CDISC-compliant ADaM dataset.

Usage

gt_adsl

Format

A tibble with 182 rows and 14 variables:

STUDYID, STUDYIDN

The unique study identifier and its numeric version.

USUBJID

The unique subject identifier.

TRTA, TRTAN

The study intervention and its numeric version, which is either Placebo (1), Drug 1 (2) or (`3`, missing for screen failures).} \item{ITTFL}{Intent-to-Treat (ITT) population flag, where `Y` indicates a subject belongs to the ITT population and `N` indicates a subject is not in the ITT population.} \item{RANDFL}{Randomization flag, where `Y` indicates a subject was randomized to receive either `Placebo` or `Drug 1` and `N` indicates a subject was not randomized at all.} \item{SCRFREAS}{The reason for screen failure. This is either missing () for non-screen failure subjects or indicates the reason for screen failure

AGE

The age of a subject at baseline in years.

AAGEGR1

The analysis age group, indicating if a subject was strictly younger than 40 years at baseline or older.

SEX

Sex of a subject. Can be either Male, Female or Undifferentiated.

ETHNIC

Ethnicity of a subject. Can be either Hispanic or Latino, Not Hispanic or Latino or missing (``).

BLBMI

Body Mass Index (BMI) of a subject at baseline in kg/m2.

EVNTFL

Event Flag. Indicates whether the subject experienced a specific event during the course of the study or not. Can be either Y, if the subject had the event, or `N`

Examples

Here is the gt_adsl.

gt_adsl
#> # A tibble: 182 x 14
#>    STUDYID STUDYIDN USUBJID TRTA  TRTAN ITTFL RANDFL SCRFR~1   AGE AAGEGR1 SEX  
#>    <chr>   <chr>    <chr>   <fct> <dbl> <chr> <chr>  <chr>   <int> <fct>   <fct>
#>  1 GT01    4001     GT1000  <NA>      3 N     N      "WITHD~    37 <40     Male 
#>  2 GT01    4001     GT1001  Plac~     1 Y     Y      ""         41 >=40    Male 
#>  3 GT01    4001     GT1002  Plac~     1 Y     Y      ""         39 <40     Fema~
#>  4 GT01    4001     GT1003  Plac~     1 Y     Y      ""         38 <40     Male 
#>  5 GT01    4001     GT1004  Plac~     1 Y     Y      ""         45 >=40    Male 
#>  6 GT01    4001     GT1005  Plac~     1 Y     Y      ""         35 <40     Fema~
#>  7 GT01    4001     GT1006  Plac~     1 Y     Y      ""         42 >=40    Fema~
#>  8 GT01    4001     GT1007  Plac~     1 Y     Y      ""         35 <40     Male 
#>  9 GT01    4001     GT1008  Plac~     1 Y     Y      ""         42 >=40    Male 
#> 10 GT01    4001     GT1009  Plac~     1 Y     Y      ""         38 <40     Fema~
#> # ... with 172 more rows, 3 more variables: ETHNIC <fct>, BLBMI <dbl>,
#> #   EVNTFL <chr>, and abbreviated variable name 1: SCRFREAS

Function ID

11-7

See also