Preview a gt table object

Create a preview version of table. Here we can specify the first and last n rows for the preview table.

gt_preview(data, rownames_to_stub = FALSE, top_n = 5, bottom_n = 1)

Arguments

data
a data.frame object or a tibble.

rownames_to_stub
an option to take rownames from the input data table as row captions in the stub.

top_n
this value will be used as the number of rows from the top of the table to display.

bottom_n
the value will be used as the number of rows from the bottom of the table to display.

Value

an object of class gt_tbl.

See also

Examples

# Create a preview of the `mtcars` dataset; it
# will be displayed in the Viewer
gt_preview(
  mtcars,
  rownames_to_stub = TRUE)

#> <style>/* gt compiled CSS file */
#> .gt_table {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#>   border-collapse: collapse;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: 100%;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> .heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#> }
#> 
#> .title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#> }
#> 
#> .headnote {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.headnote.font.size */
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#> }
#> 
#> .bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> .column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> .col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* boxhead.background.color */
#>   font-size: 16px;
#>   /* boxhead.font.size */
#>   font-weight: initial;
#>   /* boxhead.font.weight */
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> /* this is used to create a gutter between adjacent column spanners;
#>    the color should match that of .heading's background-color */
#> .sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> .group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#> }
#> 
#> .striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> .row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#> }
#> 
#> .stub {
#>   width: 10%;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#> }
#> 
#> .stub.row.right, .stub.row.left, .stub.row.center {
#>   background-color: #FFFFFF;
#> }
#> 
#> .summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> .first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> .table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> .footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> .sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> .center {
#>   text-align: center;
#> }
#> 
#> .left {
#>   text-align: left;
#> }
#> 
#> .right {
#>   text-align: right;
#> }
#> 
#> .font_normal {
#>   font-weight: normal;
#> }
#> 
#> .font_bold {
#>   font-weight: bold;
#> }
#> 
#> .font_italic {
#>   font-style: italic;
#> }
#> 
#> .gt_super {
#>   font-size: 65%;
#> }
#> </style>
#> <!--gt table start-->
#> <table data-nrow='7' data-ncol='12' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th class='col_heading right' rowspan='1' colspan='1'></th>
#> <th class='col_heading center' rowspan='1' colspan='1'>mpg</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>cyl</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>disp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>hp</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>drat</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>wt</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>qsec</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>vs</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>am</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>gear</th>
#> <th class='col_heading center' rowspan='1' colspan='1'>carb</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='stub row right'>Mazda RX4</td>
#> <td class='row center'>21</td>
#> <td class='row center'>6</td>
#> <td class='row center'>160</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.9</td>
#> <td class='row center'>2.62</td>
#> <td class='row center'>16.46</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='stub row right'>Mazda RX4 Wag</td>
#> <td class='row center'>21</td>
#> <td class='row center'>6</td>
#> <td class='row center'>160</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.9</td>
#> <td class='row center'>2.875</td>
#> <td class='row center'>17.02</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>4</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='stub row right'>Datsun 710</td>
#> <td class='row center'>22.8</td>
#> <td class='row center'>4</td>
#> <td class='row center'>108</td>
#> <td class='row center'>93</td>
#> <td class='row center'>3.85</td>
#> <td class='row center'>2.32</td>
#> <td class='row center'>18.61</td>
#> <td class='row center'>1</td>
#> <td class='row center'>1</td>
#> <td class='row center'>4</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='stub row right'>Hornet 4 Drive</td>
#> <td class='row center'>21.4</td>
#> <td class='row center'>6</td>
#> <td class='row center'>258</td>
#> <td class='row center'>110</td>
#> <td class='row center'>3.08</td>
#> <td class='row center'>3.215</td>
#> <td class='row center'>19.44</td>
#> <td class='row center'>1</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>1</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='stub row right'>Hornet Sportabout</td>
#> <td class='row center'>18.7</td>
#> <td class='row center'>8</td>
#> <td class='row center'>360</td>
#> <td class='row center'>175</td>
#> <td class='row center'>3.15</td>
#> <td class='row center'>3.44</td>
#> <td class='row center'>17.02</td>
#> <td class='row center'>0</td>
#> <td class='row center'>0</td>
#> <td class='row center'>3</td>
#> <td class='row center'>2</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='stub row right'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> <td class='row center'>...</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='stub row right'>Maserati Bora</td>
#> <td class='row center'>15</td>
#> <td class='row center'>8</td>
#> <td class='row center'>301</td>
#> <td class='row center'>335</td>
#> <td class='row center'>3.54</td>
#> <td class='row center'>3.57</td>
#> <td class='row center'>14.6</td>
#> <td class='row center'>0</td>
#> <td class='row center'>1</td>
#> <td class='row center'>5</td>
#> <td class='row center'>8</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>