info_google_fonts {gt}	R Documentation

View a table on recommended Google Fonts

Description

The google_font() helper function can be used wherever a font name should be specified. There are two instances where this helper can be used: the name argument in opt_table_font() (for setting a table font) and in that of cell_text() (used with tab_style()). Because there is an overwhelming number of fonts available in the Google Fonts catalog, the info_google_fonts() provides a table with a set of helpful font recommendations. These fonts look great in the different parts of a gt table. Why? For the most part they are suitable for body text, having large counters, large x-height, reasonably low contrast, and open apertures. These font features all make for high legibility at smaller sizes.

Usage

info_google_fonts()

Value

An object of class gt_tbl.

Function ID

10-6

See Also

Other Information Functions: info_currencies(), info_date_style(), info_locales(), info_paletteer(), info_time_style()

Examples

# Get a table of info on some of the
# recommended Google Fonts for tables
tab_1 <- info_google_fonts()

[Package gt version 0.2.2 Index]