info_locales {gt}	R Documentation

View a table with info on supported locales

Description

Many of the fmt_*() functions have a locale argument that makes locale-based formatting easier. The table generated by the info_locales() function provides a quick reference to all the available locales. The locale identifiers are provided (base locale ID, common display name) along with the each locale's group and decimal separator marks. A formatted numeric example is provided (based on the value of 11027) to demonstrate the default formatting of each locale.

Usage

info_locales(begins_with = NULL)

Arguments

begins_with
Providing a single letter will filter locales to only those that begin with that letter in their base locale ID. The default (NULL) will produce a table with all locales displayed.

Details

There are 712 locales, which means that a very long display table is provided by default. To trim down the output table size, we can provide an initial letter corresponding to the base locale ID to begins_with. This will filter locales in the info table to just the set that begins with the supplied letter.

Value

An object of class gt_tbl.

Function ID

10-4

See Also

Other Information Functions: info_currencies(), info_date_style(), info_google_fonts(), info_paletteer(), info_time_style()

Examples

# Get a table of info on all of
# the locales where the base
# locale ID begins with a "v"
tab_1 <- info_locales(begins_with = "v")

[Package gt version 0.2.2 Index]