Helpers for targeting multiple cells in different locations

These helper functions are used to target individual cells in different locations.

cells_title(groups = c("title", "subtitle"))

cells_boxhead(columns, groups)

cells_group(groups)

cells_stub(rows = NULL)

cells_data(columns = NULL, rows = NULL)

cells_summary(groups = NULL, columns = NULL, rows = NULL)

Arguments

columns, rows, groups
either a vector of names, a vector of indices, values provided by vars(), values provided by c(), or a select helper function (see Details for information on these functions).

Value

a list object of class location_cells.

Details

The following helper functions can be used to target cells (roughly in order from the top to the bottom of a table):

cells_title(): targets the table title or the table subtitle depending on the value given to the groups argument ("title" or "subtitle").

cells_boxhead(): targets captions for the column headers (the columns argument) or spanner column labels (the groups argument) in the table's column labels part.

cells_group(): targets the row group labels in any available row groups using the groups argument.

cells_stub(): targets row labels in the table stub using the rows argument.

cells_data(): targets data cells in the table field using intersections of columns and rows.

cells_summary(): targets summary cells in the table field using the groups argument and intersections of columns and rows.

The select helper functions are: starts_with(), ends_with(), contains(), matches(), one_of(), and everything().