A year of pizza sales from a pizza place

A synthetic dataset that describes pizza sales for a pizza place somewhere in the US. While the contents are artificial, the ingredients used to make the pizzas are far from it. There are 32 different pizzas that fall into 4 different categories: classic (classic pizzas: 'You probably had one like it before, but never like this!'), chicken (pizzas with chicken as a major ingredient: 'Try the Southwest Chicken Pizza! You'll love it!'), supreme (pizzas that try a little harder: 'My Soppressata pizza uses only the finest salami from my personal salumist!'), and, veggie (pizzas without any meats whatsoever: 'My Five Cheese pizza has so many cheeses, I can only offer it in Large Size!').

pizzaplace

Format

A tibble with 49574 rows and 7 variables:

id

The ID for the order, which consists of one or more pizzas at a give date and time

date

A character representation of the order date, expressed in the ISO 8601 date format (YYYY-MM-DD)

date

A character representation of the order time, expressed as a 24-hour time the ISO 8601 extended time format (hh:mm:ss)

name

The short name for the pizza

size

The size of the pizza, which can either be S, M, L, XL (rare!), or XXL (even rarer!); most pizzas are available in the S, M, and L sizes but exceptions apply

type

The category or type of pizza, which can either be classic, chicken, supreme, or veggie

price

The price of the pizza and the amount that it sold for (in USD)