px {gt}	R Documentation

Helper for providing a numeric value as pixels value

Description

Helper for providing a numeric value as pixels value

Usage

px(x)

Arguments

x
the numeric value to format as a string (e.g., "12px") for some tab_options() arguments that can take values as units of pixels (e.g., table.font.size).

See Also

Other helper functions: cells_styles, escape_latex, gt_latex_dependencies, html, md, pct

[Package gt version 0.1.0 Index]