random_id {gt}	R Documentation

Helper for creating a random id for a gt table

Description

This helper function can be used to create a random, character-based ID value argument of variable length (the default is 10 letters).

Usage

random_id(n = 10)

Arguments

n
The number of lowercase letters to use for the random ID.

Value

A character vector containing a single, random ID.

Function ID

7-25

See Also

Other Helper Functions: adjust_luminance(), cell_borders(), cell_fill(), cell_text(), cells_body(), cells_column_labels(), cells_column_spanners(), cells_footnotes(), cells_grand_summary(), cells_row_groups(), cells_source_notes(), cells_stub_grand_summary(), cells_stub_summary(), cells_stubhead(), cells_stub(), cells_summary(), cells_title(), currency(), default_fonts(), escape_latex(), google_font(), gt_latex_dependencies(), html(), md(), pct(), px()

[Package gt version 0.2.2 Index]