Reaction rates for gas-phase atmospheric reactions of organic compounds

The reactions dataset contains kinetic data for second-order (two body) gas-phase chemical reactions for 1,683 organic compounds. The reaction-rate values and parameters within this dataset are useful for studies of the atmospheric environment. Organic pollutants, which are present in trace amounts in the atmosphere, have been extensively studied by research groups since their persistence in the atmosphere requires specific attention. Many researchers have reported kinetic data on specific gas-phase reactions and these mainly involve oxidation reactions with OH, NO3 radicals, ozone, and chlorine (Cl) atoms.

This compilation of rate constant (k) data as contains the values for rate constants at 298 K (in units of cm^3 molecules^–1 s^–1) as well as parameters that allow for the calculation of rate constants at different temperatures (the temperature dependence parameters: A, B, and n). Uncertainty values/factors and temperature limits are also provided here where information is available.

Usage

reactions

Format

A tibble with 1,683 rows and 39 variables:

compd_name

The name of the primary compound undergoing reaction with OH, ozone, NO3, or Cl.

cmpd_mwt

The molecular weight of the compound in units of g/mol.

cmpd_formula

The atomic formula of the compound.

cmpd_type

The category of compounds that the compd_name falls under.

cmpd_smiles

The SMILES (simplified molecular-input line-entry system) representation for the compound.

cmpd_inchi

The InChI (International Chemical Identifier) representation for the compound.

cmpd_inchikey

The InChIKey, which is a hashed InChI value, has a fixed length of 27 characters. These values can be used to more easily perform database searches of chemical compounds.

oh_k298

Rate constant at 298 K for OH reactions.

oh_uncert

Uncertainty as a percentage for certain OH reactions.

oh_u_fac

Uncertainty as a plus/minus difference for certain OH reactions.

oh_a, oh_b, oh_n

Extended temperature dependence parameters for bimolecular OH reactions, to be used in the Arrhenius expression: k(T)=A exp(-B/T) (T/300)^n. In that, A is expressed as cm^3 molecules^-1 s^-1, B is in units of K, and n is dimensionless. Any NA values indicate that data is not available.

oh_t_low, oh_t_high

The low and high temperature boundaries (in units of K) for which the oh_a, oh_b, and oh_n parameters are valid.

o3_k298

Rate constant at 298 K for ozone reactions.

o3_uncert

Uncertainty as a percentage for certain ozone reactions.

o3_u_fac

Uncertainty as a plus/minus difference for certain ozone reactions.

o3_a, o3_b, o3_n

Extended temperature dependence parameters for bimolecular ozone reactions, to be used in the Arrhenius expression: k(T)=A exp(-B/T) (T/300)^n. In that, A is expressed as cm^3 molecules^-1 s^-1, B is in units of K, and n is dimensionless. Any NA values indicate that data is not available.

o3_t_low, o3_t_high

The low and high temperature boundaries (in units of K) for which the o3_a, o3_b, and o3_n parameters are valid.

no3_k298

Rate constant at 298 K for NO3 reactions.

no3_uncert

Uncertainty as a percentage for certain NO3 reactions.

no3_u_fac

Uncertainty as a plus/minus difference for certain NO3 reactions.

no3_a, no3_b, no3_n

Extended temperature dependence parameters for bimolecular NO3 reactions, to be used in the Arrhenius expression: k(T)=A exp(-B/T) (T/300)^n. In that, A is expressed as cm^3 molecules^-1 s^-1, B is in units of K, and n is dimensionless. Any NA values indicate that data is not available.

no3_t_low, no3_t_high

The low and high temperature boundaries (in units of K) for which the no3_a, no3_b, and no3_n parameters are valid.

cl_k298

Rate constant at 298 K for Cl reactions.

cl_uncert

Uncertainty as a percentage for certain Cl reactions.

cl_u_fac

Uncertainty as a plus/minus difference for certain Cl reactions.

cl_a, cl_b, cl_n

Extended temperature dependence parameters for bimolecular Cl reactions, to be used in the Arrhenius expression: k(T)=A exp(-B/T) (T/300)^n. In that, A is expressed as cm^3 molecules^-1 s^-1, B is in units of K, and n is dimensionless. Any NA values indicate that data is not available.

cl_t_low, cl_t_high

The low and high temperature boundaries (in units of K) for which the cl_a, cl_b, and cl_n parameters are valid.

Examples

Here is a glimpse at the data available in reactions.

dplyr::glimpse(reactions)
#> Rows: 1,683
#> Columns: 39
#> $ cmpd_name     <chr> "methane", "formaldehyde", "methanol", "fluoromethane", ~
#> $ cmpd_mwt      <dbl> 16.04, 30.03, 32.04, 34.03, 46.03, 48.02, 48.04, 50.49, ~
#> $ cmpd_formula  <chr> "CH4", "CH2O", "CH4O", "CH3F", "CH2O2", "CHOF", "CH4O2",~
#> $ cmpd_type     <chr> "normal alkane", "aldehyde", "alcohol or glycol", "haloa~
#> $ cmpd_smiles   <chr> "C", "C=O", "CO", "CF", "OC=O", "FC=O", "COO", "CCl", "F~
#> $ cmpd_inchi    <chr> "InChI=1S/CH4/h1H4", "InChI=1S/CH2O/c1-2/h1H2", "InChI=1~
#> $ cmpd_inchikey <chr> "VNWKTOKETHGBQD-UHFFFAOYSA-N", "WSFSSNUMVMOOMR-UHFFFAOYS~
#> $ oh_k298       <dbl> 6.36e-15, 8.50e-12, 8.78e-13, 1.97e-14, 4.50e-13, NA, 1.~
#> $ oh_uncert     <dbl> 0.10, 0.20, 0.10, 0.10, NA, NA, NA, 0.20, 0.10, 0.21, 0.~
#> $ oh_u_fac      <dbl> NA, NA, NA, NA, 1.4, NA, 2.0, NA, NA, NA, NA, NA, NA, NA~
#> $ oh_a          <dbl> 3.62e-13, 5.40e-12, 2.32e-13, 1.99e-13, 4.50e-13, NA, 5.~
#> $ oh_b          <dbl> 1200.3487, -135.0000, -402.0000, 685.4204, NA, NA, -190.~
#> $ oh_n          <dbl> 2.179936, NA, 2.720000, 2.040182, NA, NA, NA, NA, 1.8600~
#> $ oh_t_low      <dbl> 200, 200, 210, 240, 290, NA, 220, 220, 220, 298, 298, NA~
#> $ oh_t_high     <dbl> 2025, 300, 1344, 1800, 450, NA, 430, 330, 1800, 671, 393~
#> $ o3_k298       <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_uncert     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_u_fac      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_a          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_b          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_n          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_t_low      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ o3_t_high     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ no3_k298      <dbl> NA, 5.5e-16, 1.3e-16, NA, NA, NA, NA, NA, NA, NA, NA, NA~
#> $ no3_uncert    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ no3_u_fac     <dbl> NA, 1.6, 3.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA~
#> $ no3_a         <dbl> NA, NA, 9.4e-13, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,~
#> $ no3_b         <dbl> NA, NA, 2650, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA~
#> $ no3_n         <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~
#> $ no3_t_low     <dbl> NA, NA, 250, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,~
#> $ no3_t_high    <dbl> NA, NA, 370, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,~
#> $ cl_k298       <dbl> 1.00e-13, 7.20e-11, 5.10e-11, 3.60e-13, 1.90e-13, 1.90e-~
#> $ cl_uncert     <dbl> 0.15, 0.15, 0.20, NA, NA, 0.11, NA, 0.10, 0.15, NA, 0.14~
#> $ cl_u_fac      <dbl> NA, NA, NA, 1.4, 1.4, NA, 3.0, NA, NA, NA, NA, NA, 2.0, ~
#> $ cl_a          <dbl> 6.60e-12, 8.10e-11, 5.10e-11, 4.90e-12, NA, NA, NA, 4.32~
#> $ cl_b          <dbl> 1240.0, 34.0, 0.0, 781.0, NA, NA, NA, 646.4, 1591.0, NA,~
#> $ cl_n          <dbl> NA, NA, NA, NA, NA, NA, NA, 1.3057, NA, NA, NA, NA, NA, ~
#> $ cl_t_low      <dbl> 200, 200, 225, 200, NA, NA, NA, 222, 250, NA, NA, 220, 2~
#> $ cl_t_high     <dbl> 300, 500, 950, 300, NA, NA, NA, 843, 300, NA, NA, 330, 3~

Dataset ID and Badge

DATA-11

Dataset Introduced

in development

See also