Daily S&P 500 Index data from 1950 to 2015

This dataset provides daily price indicators for the S&P 500 index from the beginning of 1950 to the end of 2015. The index includes 500 leading companies and captures about 80\

sp500

Format

A tibble with 16607 rows and 7 variables:

date

The date expressed as Date values

open, high, low, close

The day's opening, high, low, and closing prices in USD; the close price is adjusted for splits

volume

the number of trades for the given date

adj_close

The close price adjusted for both dividends and splits

Function ID

11-4

See also

Examples

# Here is a glimpse at the data
# available in `sp500`
dplyr::glimpse(sp500)

#> Rows: 16,607
#> Columns: 7
#> $ date      <date> 2015-12-31, 2015-12-30, 2015-12-29, 2015-12-28, 2015-12-24…
#> $ open      <dbl> 2060.59, 2077.34, 2060.54, 2057.77, 2063.52, 2042.20, 2023.…
#> $ high      <dbl> 2062.54, 2077.34, 2081.56, 2057.77, 2067.36, 2064.73, 2042.…
#> $ low       <dbl> 2043.62, 2061.97, 2060.54, 2044.20, 2058.73, 2042.20, 2020.…
#> $ close     <dbl> 2043.94, 2063.36, 2078.36, 2056.50, 2060.99, 2064.29, 2038.…
#> $ volume    <dbl> 2655330000, 2367430000, 2542000000, 2492510000, 1411860000,…
#> $ adj_close <dbl> 2043.94, 2063.36, 2078.36, 2056.50, 2060.99, 2064.29, 2038.…