Add summary rows using aggregation functions

Add summary rows to one or more stub blocks by using the input data already provided in the gt() function alongside any suitable aggregation functions. Should we need to obtain the summary data for external purposes, the extract_summary() can be used with a gt_tbl object where summary rows were added via summary_rows().

summary_rows(data, groups = NULL, columns = NULL, funs,
  missing_text = "---", formatter = fmt_number, ...)

Arguments

data
a table object that is created using the gt() function.

groups
the stub block groups heading names for which summary rows will be added.

columns
the columns for which the summaries should be calculated. If nothing is provided, then the supplied aggregation functions will be applied to all columns.

funs
functions used for aggregations. This can include base functions like mean, min, max, median, sd, or sum. The functions should be supplied using the funs() helper function. With funs(), we can specify the functions by use of function names (e.g., "sum"), the functions themselves (e.g., sum), or calls to functions with . as a dummy argument (e.g., sum(., na.rm = TRUE)).

missing_text
the text to be used in place of NA values in summary cells with no data outputs.

formatter
a formatter function name. These can be any of the fmt_*functions available in the package (e.g., fmt_number, link{fmt_percent}, etc.), or a custom function using fmt(). The default function is fmt_number() and its options can be accessed through ....

...
values passed to the formatter function, where the provided values are to be in the form of named vectors. For example, when using the default formatter function, fmt_number, options such as decimals, use_seps, and locale can be used.

Value

an object of class gt_tbl.

Examples

library(tidyverse)

# Create a table that creates a stub and
# stub blocks based on a naming convention
tbl <-
  dplyr::tribble(
    ~groupname, ~rowname, ~value_1, ~value_2,
    "A",        "1",      NA,       260.1,
    "A",        "2",      184.3,    84.4,
    "A",        "3",      342.3,    126.3,
    "A",        "4",      234.9,    NA,
    "B",        "1",      190.9,    832.5,
    "B",        "2",      743.3,    281.2,
    "B",        "3",      252.3,    732.5,
    "B",        "4",      344.7,    NA,
    "C",        "1",      197.2,    818.0,
    "C",        "2",      284.3,    394.4)

# Create a table with summary rows for the
# `A` and `C` groups; the 3 summary rows
# for these groups represent the mean, sum,
# and standard deviation of `value`
gt_tbl <-
  gt(tbl) %>%
    summary_rows(
      groups = c("A", "C"),
      columns = vars(value_1),
      funs = funs(
        average = mean(., na.rm = TRUE),
        total = sum(., na.rm = TRUE),
        `std dev` = sd(., na.rm = TRUE)))