Arrange a boxhead into panels

Set a spanner with a name and mappings to columns extant in the table. This creates a boxhead panel with spanner headings and column headings.

tab_boxhead_panel(data, group, columns)

Arguments

data
a table object that is created using the gt() function.

group
the name to assign to the spanner heading.

columns
the columns to be components of the spanner heading.

Value

an object of class gt_tbl.

See also

Examples

# Create a table based on `rock` where
# there are column headings grouped
# under spanner headings
gt(data = rock) %>%
  tab_boxhead_panel(
    group = "perimeter",
    columns = c("peri", "shape"))

#> <style>/* gt compiled CSS file */
#> .gt_table {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
#>   border-collapse: collapse;
#>   color: #000000;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   /* table.background.color */
#>   width: 100%;
#>   /* table.width */
#>   border-top-style: solid;
#>   /* table.border.top.style */
#>   border-top-width: 2px;
#>   /* table.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* table.border.top.color */
#> }
#> 
#> .heading {
#>   background-color: #FFFFFF;
#>   /* heading.background.color */
#> }
#> 
#> .title {
#>   color: #000000;
#>   font-size: 125%;
#>   /* heading.title.font.size */
#>   padding-top: 4px;
#>   /* heading.top.padding */
#> }
#> 
#> .headnote {
#>   color: #000000;
#>   font-size: 85%;
#>   /* heading.headnote.font.size */
#>   padding-bottom: 4px;
#>   /* heading.bottom.padding */
#> }
#> 
#> .bottom_border {
#>   border-bottom-style: solid;
#>   /* heading.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* heading.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* heading.border.bottom.color */
#> }
#> 
#> .column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> .col_heading {
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* boxhead.background.color */
#>   font-size: 16px;
#>   /* boxhead.font.size */
#>   font-weight: initial;
#>   /* boxhead.font.weight */
#>   padding: 10px;
#>   margin: 10px;
#> }
#> 
#> /* this is used to create a gutter between adjacent column spanners;
#>    the color should match that of .heading's background-color */
#> .sep_right {
#>   border-right: 5px solid #FFFFFF;
#> }
#> 
#> .group_heading {
#>   padding: 8px;
#>   color: #000000;
#>   background-color: #FFFFFF;
#>   /* stub_group.background.color */
#>   font-size: 16px;
#>   /* stub_group.font.size */
#>   font-weight: initial;
#>   /* stub_group.font.weight */
#>   border-top-style: solid;
#>   /* stub_group.border.top.style */
#>   border-top-width: 2px;
#>   /* stub_group.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* stub_group.border.top.color */
#>   border-bottom-style: solid;
#>   /* stub_group.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* stub_group.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* stub_group.border.bottom.color */
#> }
#> 
#> .striped tr:nth-child(even) {
#>   background-color: #f2f2f2;
#> }
#> 
#> .row {
#>   padding: 10px;
#>   /* row.padding */
#>   margin: 10px;
#> }
#> 
#> .stub {
#>   width: 10%;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #A8A8A8;
#> }
#> 
#> .stub.row.right, .stub.row.left, .stub.row.center {
#>   background-color: #FFFFFF;
#> }
#> 
#> .summary_row {
#>   background-color: #FFFFFF;
#>   /* summary_row.background.color */
#>   padding: 6px;
#>   /* summary_row.padding */
#>   text-transform: inherit;
#>   /* summary_row.text_transform */
#> }
#> 
#> .first_summary_row {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #A8A8A8;
#> }
#> 
#> .table_body {
#>   border-top-style: solid;
#>   /* field.border.top.style */
#>   border-top-width: 2px;
#>   /* field.border.top.width */
#>   border-top-color: #A8A8A8;
#>   /* field.border.top.color */
#>   border-bottom-style: solid;
#>   /* field.border.bottom.style */
#>   border-bottom-width: 2px;
#>   /* field.border.bottom.width */
#>   border-bottom-color: #A8A8A8;
#>   /* field.border.bottom.color */
#> }
#> 
#> .footnote {
#>   font-size: 90%;
#>   /* footnote.font.size */
#>   padding: 4px;
#>   /* footnote.padding */
#> }
#> 
#> .sourcenote {
#>   font-size: 90%;
#>   /* sourcenote.font.size */
#>   padding: 4px;
#>   /* sourcenote.padding */
#> }
#> 
#> .center {
#>   text-align: center;
#> }
#> 
#> .left {
#>   text-align: left;
#> }
#> 
#> .right {
#>   text-align: right;
#> }
#> 
#> .font_normal {
#>   font-weight: normal;
#> }
#> 
#> .font_bold {
#>   font-weight: bold;
#> }
#> 
#> .font_italic {
#>   font-style: italic;
#> }
#> 
#> .gt_super {
#>   font-size: 65%;
#> }
#> </style>
#> <!--gt table start-->
#> <table data-nrow='48' data-ncol='4' data-ngroup='0' class='gt_table'>
#> <tr>
#> <th data-type='column_heading' class='col_heading center' rowspan='2' colspan='1'>area</th>
#> <th data-type='column_heading' class='col_heading column_spanner center' rowspan='1' colspan='2'>perimeter</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='2' colspan='1'>perm</th>
#> </tr>
#> <tr>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>peri</th>
#> <th data-type='column_heading' class='col_heading center' rowspan='1' colspan='1'>shape</th>
#> </tr>
#> <tbody class='table_body striped'>
#> <tr data-type='data' data-row='1'>
#> <td class='row center' style=''>4990</td>
#> <td class='row center' style=''>2791.9</td>
#> <td class='row center' style=''>0.0903296</td>
#> <td class='row center' style=''>6.3</td>
#> </tr>
#> <tr data-type='data' data-row='2'>
#> <td class='row center' style=''>7002</td>
#> <td class='row center' style=''>3892.6</td>
#> <td class='row center' style=''>0.148622</td>
#> <td class='row center' style=''>6.3</td>
#> </tr>
#> <tr data-type='data' data-row='3'>
#> <td class='row center' style=''>7558</td>
#> <td class='row center' style=''>3930.66</td>
#> <td class='row center' style=''>0.183312</td>
#> <td class='row center' style=''>6.3</td>
#> </tr>
#> <tr data-type='data' data-row='4'>
#> <td class='row center' style=''>7352</td>
#> <td class='row center' style=''>3869.32</td>
#> <td class='row center' style=''>0.117063</td>
#> <td class='row center' style=''>6.3</td>
#> </tr>
#> <tr data-type='data' data-row='5'>
#> <td class='row center' style=''>7943</td>
#> <td class='row center' style=''>3948.54</td>
#> <td class='row center' style=''>0.122417</td>
#> <td class='row center' style=''>17.1</td>
#> </tr>
#> <tr data-type='data' data-row='6'>
#> <td class='row center' style=''>7979</td>
#> <td class='row center' style=''>4010.15</td>
#> <td class='row center' style=''>0.167045</td>
#> <td class='row center' style=''>17.1</td>
#> </tr>
#> <tr data-type='data' data-row='7'>
#> <td class='row center' style=''>9333</td>
#> <td class='row center' style=''>4345.75</td>
#> <td class='row center' style=''>0.189651</td>
#> <td class='row center' style=''>17.1</td>
#> </tr>
#> <tr data-type='data' data-row='8'>
#> <td class='row center' style=''>8209</td>
#> <td class='row center' style=''>4344.75</td>
#> <td class='row center' style=''>0.164127</td>
#> <td class='row center' style=''>17.1</td>
#> </tr>
#> <tr data-type='data' data-row='9'>
#> <td class='row center' style=''>8393</td>
#> <td class='row center' style=''>3682.04</td>
#> <td class='row center' style=''>0.203654</td>
#> <td class='row center' style=''>119</td>
#> </tr>
#> <tr data-type='data' data-row='10'>
#> <td class='row center' style=''>6425</td>
#> <td class='row center' style=''>3098.65</td>
#> <td class='row center' style=''>0.162394</td>
#> <td class='row center' style=''>119</td>
#> </tr>
#> <tr data-type='data' data-row='11'>
#> <td class='row center' style=''>9364</td>
#> <td class='row center' style=''>4480.05</td>
#> <td class='row center' style=''>0.150944</td>
#> <td class='row center' style=''>119</td>
#> </tr>
#> <tr data-type='data' data-row='12'>
#> <td class='row center' style=''>8624</td>
#> <td class='row center' style=''>3986.24</td>
#> <td class='row center' style=''>0.148141</td>
#> <td class='row center' style=''>119</td>
#> </tr>
#> <tr data-type='data' data-row='13'>
#> <td class='row center' style=''>10651</td>
#> <td class='row center' style=''>4036.54</td>
#> <td class='row center' style=''>0.228595</td>
#> <td class='row center' style=''>82.4</td>
#> </tr>
#> <tr data-type='data' data-row='14'>
#> <td class='row center' style=''>8868</td>
#> <td class='row center' style=''>3518.04</td>
#> <td class='row center' style=''>0.231623</td>
#> <td class='row center' style=''>82.4</td>
#> </tr>
#> <tr data-type='data' data-row='15'>
#> <td class='row center' style=''>9417</td>
#> <td class='row center' style=''>3999.37</td>
#> <td class='row center' style=''>0.172567</td>
#> <td class='row center' style=''>82.4</td>
#> </tr>
#> <tr data-type='data' data-row='16'>
#> <td class='row center' style=''>8874</td>
#> <td class='row center' style=''>3629.07</td>
#> <td class='row center' style=''>0.153481</td>
#> <td class='row center' style=''>82.4</td>
#> </tr>
#> <tr data-type='data' data-row='17'>
#> <td class='row center' style=''>10962</td>
#> <td class='row center' style=''>4608.66</td>
#> <td class='row center' style=''>0.204314</td>
#> <td class='row center' style=''>58.6</td>
#> </tr>
#> <tr data-type='data' data-row='18'>
#> <td class='row center' style=''>10743</td>
#> <td class='row center' style=''>4787.62</td>
#> <td class='row center' style=''>0.262727</td>
#> <td class='row center' style=''>58.6</td>
#> </tr>
#> <tr data-type='data' data-row='19'>
#> <td class='row center' style=''>11878</td>
#> <td class='row center' style=''>4864.22</td>
#> <td class='row center' style=''>0.200071</td>
#> <td class='row center' style=''>58.6</td>
#> </tr>
#> <tr data-type='data' data-row='20'>
#> <td class='row center' style=''>9867</td>
#> <td class='row center' style=''>4479.41</td>
#> <td class='row center' style=''>0.14481</td>
#> <td class='row center' style=''>58.6</td>
#> </tr>
#> <tr data-type='data' data-row='21'>
#> <td class='row center' style=''>7838</td>
#> <td class='row center' style=''>3428.74</td>
#> <td class='row center' style=''>0.113852</td>
#> <td class='row center' style=''>142</td>
#> </tr>
#> <tr data-type='data' data-row='22'>
#> <td class='row center' style=''>11876</td>
#> <td class='row center' style=''>4353.14</td>
#> <td class='row center' style=''>0.291029</td>
#> <td class='row center' style=''>142</td>
#> </tr>
#> <tr data-type='data' data-row='23'>
#> <td class='row center' style=''>12212</td>
#> <td class='row center' style=''>4697.65</td>
#> <td class='row center' style=''>0.240077</td>
#> <td class='row center' style=''>142</td>
#> </tr>
#> <tr data-type='data' data-row='24'>
#> <td class='row center' style=''>8233</td>
#> <td class='row center' style=''>3518.44</td>
#> <td class='row center' style=''>0.161865</td>
#> <td class='row center' style=''>142</td>
#> </tr>
#> <tr data-type='data' data-row='25'>
#> <td class='row center' style=''>6360</td>
#> <td class='row center' style=''>1977.39</td>
#> <td class='row center' style=''>0.280887</td>
#> <td class='row center' style=''>740</td>
#> </tr>
#> <tr data-type='data' data-row='26'>
#> <td class='row center' style=''>4193</td>
#> <td class='row center' style=''>1379.35</td>
#> <td class='row center' style=''>0.179455</td>
#> <td class='row center' style=''>740</td>
#> </tr>
#> <tr data-type='data' data-row='27'>
#> <td class='row center' style=''>7416</td>
#> <td class='row center' style=''>1916.24</td>
#> <td class='row center' style=''>0.191802</td>
#> <td class='row center' style=''>740</td>
#> </tr>
#> <tr data-type='data' data-row='28'>
#> <td class='row center' style=''>5246</td>
#> <td class='row center' style=''>1585.42</td>
#> <td class='row center' style=''>0.133083</td>
#> <td class='row center' style=''>740</td>
#> </tr>
#> <tr data-type='data' data-row='29'>
#> <td class='row center' style=''>6509</td>
#> <td class='row center' style=''>1851.21</td>
#> <td class='row center' style=''>0.225214</td>
#> <td class='row center' style=''>890</td>
#> </tr>
#> <tr data-type='data' data-row='30'>
#> <td class='row center' style=''>4895</td>
#> <td class='row center' style=''>1239.66</td>
#> <td class='row center' style=''>0.341273</td>
#> <td class='row center' style=''>890</td>
#> </tr>
#> <tr data-type='data' data-row='31'>
#> <td class='row center' style=''>6775</td>
#> <td class='row center' style=''>1728.14</td>
#> <td class='row center' style=''>0.311646</td>
#> <td class='row center' style=''>890</td>
#> </tr>
#> <tr data-type='data' data-row='32'>
#> <td class='row center' style=''>7894</td>
#> <td class='row center' style=''>1461.06</td>
#> <td class='row center' style=''>0.276016</td>
#> <td class='row center' style=''>890</td>
#> </tr>
#> <tr data-type='data' data-row='33'>
#> <td class='row center' style=''>5980</td>
#> <td class='row center' style=''>1426.76</td>
#> <td class='row center' style=''>0.197653</td>
#> <td class='row center' style=''>950</td>
#> </tr>
#> <tr data-type='data' data-row='34'>
#> <td class='row center' style=''>5318</td>
#> <td class='row center' style=''>990.388</td>
#> <td class='row center' style=''>0.326635</td>
#> <td class='row center' style=''>950</td>
#> </tr>
#> <tr data-type='data' data-row='35'>
#> <td class='row center' style=''>7392</td>
#> <td class='row center' style=''>1350.76</td>
#> <td class='row center' style=''>0.154192</td>
#> <td class='row center' style=''>950</td>
#> </tr>
#> <tr data-type='data' data-row='36'>
#> <td class='row center' style=''>7894</td>
#> <td class='row center' style=''>1461.06</td>
#> <td class='row center' style=''>0.276016</td>
#> <td class='row center' style=''>950</td>
#> </tr>
#> <tr data-type='data' data-row='37'>
#> <td class='row center' style=''>3469</td>
#> <td class='row center' style=''>1376.7</td>
#> <td class='row center' style=''>0.176969</td>
#> <td class='row center' style=''>100</td>
#> </tr>
#> <tr data-type='data' data-row='38'>
#> <td class='row center' style=''>1468</td>
#> <td class='row center' style=''>476.322</td>
#> <td class='row center' style=''>0.438712</td>
#> <td class='row center' style=''>100</td>
#> </tr>
#> <tr data-type='data' data-row='39'>
#> <td class='row center' style=''>3524</td>
#> <td class='row center' style=''>1189.46</td>
#> <td class='row center' style=''>0.163586</td>
#> <td class='row center' style=''>100</td>
#> </tr>
#> <tr data-type='data' data-row='40'>
#> <td class='row center' style=''>5267</td>
#> <td class='row center' style=''>1644.96</td>
#> <td class='row center' style=''>0.253832</td>
#> <td class='row center' style=''>100</td>
#> </tr>
#> <tr data-type='data' data-row='41'>
#> <td class='row center' style=''>5048</td>
#> <td class='row center' style=''>941.543</td>
#> <td class='row center' style=''>0.328641</td>
#> <td class='row center' style=''>1300</td>
#> </tr>
#> <tr data-type='data' data-row='42'>
#> <td class='row center' style=''>1016</td>
#> <td class='row center' style=''>308.642</td>
#> <td class='row center' style=''>0.230081</td>
#> <td class='row center' style=''>1300</td>
#> </tr>
#> <tr data-type='data' data-row='43'>
#> <td class='row center' style=''>5605</td>
#> <td class='row center' style=''>1145.69</td>
#> <td class='row center' style=''>0.464125</td>
#> <td class='row center' style=''>1300</td>
#> </tr>
#> <tr data-type='data' data-row='44'>
#> <td class='row center' style=''>8793</td>
#> <td class='row center' style=''>2280.49</td>
#> <td class='row center' style=''>0.420477</td>
#> <td class='row center' style=''>1300</td>
#> </tr>
#> <tr data-type='data' data-row='45'>
#> <td class='row center' style=''>3475</td>
#> <td class='row center' style=''>1174.11</td>
#> <td class='row center' style=''>0.200744</td>
#> <td class='row center' style=''>580</td>
#> </tr>
#> <tr data-type='data' data-row='46'>
#> <td class='row center' style=''>1651</td>
#> <td class='row center' style=''>597.808</td>
#> <td class='row center' style=''>0.262651</td>
#> <td class='row center' style=''>580</td>
#> </tr>
#> <tr data-type='data' data-row='47'>
#> <td class='row center' style=''>5514</td>
#> <td class='row center' style=''>1455.88</td>
#> <td class='row center' style=''>0.182453</td>
#> <td class='row center' style=''>580</td>
#> </tr>
#> <tr data-type='data' data-row='48'>
#> <td class='row center' style=''>9718</td>
#> <td class='row center' style=''>1485.58</td>
#> <td class='row center' style=''>0.200447</td>
#> <td class='row center' style=''>580</td>
#> </tr>
#> </tbody>
#> </table>
#> <!--gt table end-->
#>