Modify the table output options

Modify the options available in a table. These options are named by the components, the subcomponents, and the element that can adjusted.

tab_options(data, table.width = NULL, table.font.size = NULL,
  table.background.color = NULL, table.border.top.style = NULL,
  table.border.top.width = NULL, table.border.top.color = NULL,
  heading.background.color = NULL, heading.title.font.size = NULL,
  heading.headnote.font.size = NULL,
  heading.border.bottom.style = NULL,
  heading.border.bottom.width = NULL,
  heading.border.bottom.color = NULL, boxhead.background.color = NULL,
  boxhead.font.size = NULL, boxhead.font.weight = NULL,
  stub_group.background.color = NULL, stub_group.font.size = NULL,
  stub_group.font.weight = NULL, stub_group.border.top.style = NULL,
  stub_group.border.top.width = NULL,
  stub_group.border.top.color = NULL,
  stub_group.border.bottom.style = NULL,
  stub_group.border.bottom.width = NULL,
  stub_group.border.bottom.color = NULL, field.border.top.style = NULL,
  field.border.top.width = NULL, field.border.top.color = NULL,
  field.border.bottom.style = NULL, field.border.bottom.width = NULL,
  field.border.bottom.color = NULL, row.padding = NULL,
  summary_row.background.color = NULL, summary_row.padding = NULL,
  summary_row.text_transform = NULL, footnote.sep = NULL,
  footnote.glyph = NULL, footnote.font.size = NULL,
  footnote.padding = NULL, sourcenote.font.size = NULL,
  sourcenote.padding = NULL, row.striping.include_stub = NULL,
  row.striping.include_field = NULL)

Arguments

data
a table object that is created using the gt() function.

table.width
the width of the table. Can be specified as a single-length character with units of pixels or as a percentage. If provided as a single-length numeric vector, it is assumed that the value is given in units of pixels. The px() and pct() helper functions can also be used to pass in numeric values and obtain values as pixel or percent units.

table.font.size, heading.title.font.size, heading.headnote.font.size, boxhead.font.size, stub_group.font.size, footnote.font.size, sourcenote.font.size
font sizes for the parent text element table and the following child elements: heading.title, heading.headnote, boxhead, stub_group, footnote, and sourcenote. Can be specified as a single-length character vector with units of pixels (e.g., 12px) or as a percentage (e.g., 80%). If provided as a single-length numeric vector, it is assumed that the value is given in units of pixels. The px() and pct() helper functions can also be used to pass in numeric values and obtain values as pixel or percent units.

table.background.color, heading.background.color, boxhead.background.color, stub_group.background.color, summary_row.background.color
background colors for the parent element table and the following child elements: heading, boxhead, stub_group, summary_row, and field. A color name or a hexadecimal color code should be provided.

table.border.top.style, table.border.top.width, table.border.top.color
the style, width, and color of the table's top border.

heading.border.bottom.style, heading.border.bottom.width, heading.border.bottom.color
the style, width, and color of the heading's bottom border.

boxhead.font.weight, stub_group.font.weight
the font weight of the boxhead and stub_group text element.

stub_group.border.top.style, stub_group.border.top.width, stub_group.border.top.color
the style, width, and color of the stub heading's top border.

stub_group.border.bottom.style, stub_group.border.bottom.width, stub_group.border.bottom.color
the style, width, and color of the stub heading's bottom border.

field.border.top.style, field.border.top.width, field.border.top.color
the style, width, and color of the field's top border.

field.border.bottom.style, field.border.bottom.width, field.border.bottom.color
the style, width, and color of the field's bottom border.

row.padding, summary_row.padding
the amount of padding in each row and in each summary row.

summary_row.text_transform
an option to apply text transformations to the label text in each summary row.

footnote.sep
the separating characters between adjacent footnotes in the footnotes section. The default value produces a linebreak.

footnote.glyph
the set of sequential figures or characters used to identify the footnotes. We can either supply the keyword "numbers" (the default, indicating that we want numeric glyphs), the keywords "letters" or "LETTERS" (indicating that we want letters as glyphs, either lowercase or uppercase), or, a vector of character values representing the series of glyphs. A series of glyphs is recycled when its usage goes beyond the length of the set. At each cycle, the glyphs are simply combined (e.g., * -> ** -> ***).

footnote.padding, sourcenote.padding
the amount of padding to apply to the footnote and source note sections.

row.striping.include_stub
an option for whether to include the stub when striping rows.

row.striping.include_field
an option for whether to include the field when striping rows.

See also