Arrange a stub into blocks

Set a group with a name and mappings to rows extant in the table. This creates a stub block with group headings and row captions.

tab_stub_block(data, group = NULL, rows = NULL, others = NULL)

Arguments

data
a table object that is created using the gt() function.

group
the stub block group heading name.

rows
the rows to be made components of the stub block. Can either be a vector of row captions provided c(), a vector of row indices, or a helper function focused on selections. The select helper functions are: starts_with(), ends_with(), contains(), matches(), one_of(), and everything().

others
an option to set a default group heading for any rows not formally placed in a stub block named by group in any call of tab_stub_block(). A separate call to tab_stub_block() with only a value to others is possible and makes explicit that the call is meant to provide a default group heading value. If this is not set and there are rows that haven't been placed into a stub block (where one or more stub blocks exist), then those rows will be automatically placed into a stub block without a title.

Value

an object of class gt_tbl.

See also

Examples

# Create a table based on `mtcars` where
# there are group headings grouped inside
# stub blocks; two groups are generated
# here: `Mazda` and `NA` (a group without)
# a title
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) %>%
    tab_stub_block(
      group = "Mazda",
      rows = c("Mazda RX4", "Mazda RX4 Wag"))