Add label text to the stubhead

Add a label to the stubhead of a table. If a stub does not exist, no change will be made.

tab_stubhead_label(data, label)

Arguments

data
a table object that is created using the gt() function.

label
the text to be used as the stubhead label We can optionally use the md() and html() functions to style the text as Markdown or to retain HTML elements in the text.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `mtcars` dataset and add a stubhead
# label to describe what is in the stub
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) %>%
    tab_stubhead_label(
      label = md("car *make* and *model*"))