test_image {gt}	R Documentation

Generate a path to a test image

Description

Two test images are available within the gt package. Both contain the same imagery (sized at 200px by 200px) but one is a PNG file while the other is an SVG file. This function is most useful when paired with local_image() since we test various sizes of the test image within that function.

Usage

test_image(type = c("png", "svg"))

Arguments

type
the type of the image, which can either be png (the default) or svg.

See Also

Other image addition functions: ggplot_image, local_image, web_image

[Package gt version 0.1.0 Index]