Helper for defining custom styles for table cells

This helper function is to be used with the tab_style() function, which itself allows for the setting of custom styles to one or more cells. We can also define several styles with a single call of cells_styles and tab_style() will reliably process that.

cells_styles(bkgd_color = NULL, text_color = NULL, text_font = NULL,
  text_size = NULL, text_align = NULL, text_style = NULL,
  text_weight = NULL, text_stretch = NULL, text_indent = NULL,
  text_decorate = NULL, text_transform = NULL)

Arguments

bkgd_color
the background color of the cell.

text_color
the text color.

text_font
the font or collection of fonts (subsequent font names are) used as fallbacks.

text_size
the size of the font. Can either

text_align
the text alignment. Can be one of either "center", "left", "right", or "justify".

text_style
the text style. Can be one of either "center", "normal", "italic", or "oblique".

text_weight
the weight of the font. Can be a text-based keyword such as "normal", "bold", "lighter", "bolder", or, a numeric value between 1 and 1000, inclusive. Note that only variable fonts may support the numeric mapping of weight.

text_stretch
allows for text to either be condensed or expanded. We can use the following text-based keywords to describe the degree of condensation/expansion: ultra-condensed, extra-condensed, condensed, semi-condensed, normal, semi-expanded, expanded, extra-expanded, and ultra-expanded. Alternatively, we can supply percentage values from 0% to 200%, inclusive. Negative percentage values are not allowed.

text_indent
the indentation of the text.

text_decorate
allows for text decoration effect to be applied. Here, we can use "overline", "line-through", or "underline".

text_transform
allows for the transformation of text. Options are "uppercase", "lowercase", or "capitalize".

Value

a character vector containing formatted styles.

See also