A toy example tibble for testing with gt: exibble

This tibble contains data of a few different classes, which makes it well-suited for quick experimentation with the functions in this package. It contains only eight rows with numeric, character, and factor columns. The last 4 rows contain NA values in the majority of this tibble's columns (1 missing value per column). The date, time, and datetime columns are character-based dates/times in the familiar ISO 8601 format. The row and group columns provide for unique rownames and two groups (grp_a and grp_b) for experimenting with the gt() function's rowname_col and groupname_col arguments.

exibble

Format

A tibble with 8 rows and 9 variables:

num

a numeric column ordered with increasingly larger values

char

a character column composed of names of fruits from a to h

fctr

a factor column with numbers from 1 to 8, written out

date, time, datetime

character columns with dates, times, and datetimes

currency

a numeric column that is useful for testing currency-based formatting

row

a character column in the format row_X which can be useful for testing with row captions in a table stub

group

a character column with four grp_a values and four grp_b values which can be useful for testing tables that contain row groups

Function ID

11-6

See also

Examples

# Here is a glimpse at the data
# available in `exibble`
dplyr::glimpse(exibble)
#> Rows: 8
#> Columns: 9
#> $ num      <dbl> 1.111e-01, 2.222e+00, 3.333e+01, 4.444e+02, 5.550e+03, NA, 7.…
#> $ char     <chr> "apricot", "banana", "coconut", "durian", NA, "fig", "grapefr…
#> $ fctr     <fct> one, two, three, four, five, six, seven, eight
#> $ date     <chr> "2015-01-15", "2015-02-15", "2015-03-15", "2015-04-15", "2015…
#> $ time     <chr> "13:35", "14:40", "15:45", "16:50", "17:55", NA, "19:10", "20…
#> $ datetime <chr> "2018-01-01 02:22", "2018-02-02 14:33", "2018-03-03 03:44", "…
#> $ currency <dbl> 49.950, 17.950, 1.390, 65100.000, 1325.810, 13.255, NA, 0.440
#> $ row      <chr> "row_1", "row_2", "row_3", "row_4", "row_5", "row_6", "row_7"…
#> $ group    <chr> "grp_a", "grp_a", "grp_a", "grp_a", "grp_b", "grp_b", "grp_b"…