Format numeric values

Format numeric values

fmt_number(data, columns, rows = NULL, decimals = 2,
  drop_trailing_zeros = FALSE, negative_val = "signed",
  use_seps = TRUE, scale_by = 1, pattern = "{x}", sep_mark = ",",
  dec_mark = ".", locale = NULL)

Arguments

data
a table object that is created using the gt() function.

columns
the column names to format.

rows
optional rows to format. Not providing any value results in all rows in columns being formatted.

decimals
an option to specify the exact number of decimal places to use. The default number of decimal places is 2.

drop_trailing_zeros
a logical value that allows for removal of trailing zeros (those redundant zeros after the decimal mark).

negative_val
the formatting to use for negative values. With signed (the default), negative values will be shown with a negative sign. Using parens will show the negative value in parentheses.

use_seps
an option to use digit group separators. The type of digit group separator is set by sep_mark and overridden if a locale ID is provided to locale. This setting is TRUE by default.

scale_by
a value to scale the input. The default is 1.0.

pattern
a formatting pattern that allows for decoration of the formatted value. The value itself is represented by {x} and all other characters are taken to be string literals.

sep_mark
the mark to use as a separator between groups of digits.

dec_mark
the character to use as a decimal mark.

locale
an optional locale ID that can be used for formatting the value according the locale's rules. Examples include "en_US" for English (United States) and "fr_FR" for French (France). The use of a valid locale ID will override any values provided in sep_mark and dec_mark.

Value

an object of class gt_tbl.

See also

Examples

library(tidyverse)

# Create a table object using the
# `mtcars` dataset and format specified
# numeric columns to display values to
# two decimal places
gt_tbl_1 <-
  gt(mtcars, rownames_to_stub = TRUE) %>%
    fmt_number(
      columns = vars(mpg, disp, drat, qsec),
      decimals = 5)

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_1,   ~val_2,
    8236625,  -2345,
    6533405,  -7865,
    7363931,  2345,
    9326440,  7543)

# Create a gt table from `data_tbl` and
# scale the numeric values in columns
# `val_1` and `val_2` and apply a suffix
gt_tbl_2 <-
  gt(data_tbl) %>%
    fmt_number(
      columns = vars(val_1),
      decimals = 2,
      scale_by = 1E-6,
      pattern = "{x}") %>%
    fmt_number(
      columns = vars(val_2),
      decimals = 2,
      scale_by = 1/1000,
      pattern = "{x}K")