Create a gt table object

Create a gt table object so that we can perform styling transformations before transforming the entire object to a display table. Using this function is the first step in that process, and, there are a few data ingest options we can consider at this stage.

gt(data, rowname_col = "rowname", groupname_col = "groupname",
  rownames_to_stub = FALSE,
  stub_group.sep = getOption("gt.stub_group.sep", " - "))

Arguments

data
a data.frame object or a tibble.

rowname_col
the column name in the input data table to use as row captions to be placed in the display table stub.

groupname_col
the column name in the input data table to use as group labels for generation of stub row groups.

rownames_to_stub
an option to take rownames from the input data table as row captions in the display table stub.

stub_group.sep
the separator to use between consecutive group names (a possibility when providing data as a grouped_df with multiple groups) in the displayed stub row group label.

Value

an object of class gt_tbl.

See also

Examples

# Create a table object using the
# `mtcars` dataset; take the rownames
# available in this table and use them
# as row captions in the table stub
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE)

# The resulting object can be used
# in transformations (with `tab_*()`,
# `fmt_*()`, `cols_*()` functions)
gt_tbl_2 <-
  gt_tbl %>%
  tab_header(title = "mtcars") %>%
  fmt_number(
    columns = vars(drat, wt, qsec),
    decimals = 1
  ) %>%
  cols_label(hp = "HP")