gt-package {gt}	R Documentation

gt: Easily Create Presentation-Ready Display Tables

Description

Build display tables from tabular data with an easy-to-use set of functions. With its progressive approach, we can construct display tables with a cohesive set of table parts. Table values can be formatted using any of the included formatting functions. Footnotes and cell styles can be precisely added through a location targeting system. The way in which 'gt' handles things for you means that you don't often have to worry about the fine details.

Author(s)

Maintainer: Richard Iannone rich@rstudio.com (ORCID)

Authors:

Joe Cheng joe@rstudio.com

Barret Schloerke barret@rstudio.com (ORCID)

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://gt.rstudio.com/

https://github.com/rstudio/gt

Report bugs at https://github.com/rstudio/gt/issues

[Package gt version 0.2.2 Index]