gt: Easily Create Presentation-Ready Display Tables

Build display tables from tabular data with an easy-to-use set of functions. With its progressive approach, we can construct display tables with a cohesive set of table parts. Table values can be formatted using any of the included formatter functions. Footnotes and cell styles can be precisely added through gt's location targeting system. The way in which gt handles things for you means that you don't often have to worry about the fine details.

See also

Author

Maintainer: Richard Iannone rich@rstudio.com (ORCID)

Authors:

Joe Cheng joe@rstudio.com

Barret Schloerke barret@rstudio.com (ORCID)

Other contributors:

RStudio [copyright holder, funder]