gt: Easily Create Presentation-Ready Display Tables

Build display tables from tabular data with an easy-to-use set of functions. With its progressive approach, we can construct display tables with a cohesive set of table parts. Table values can be formatted using any of the included formatting functions. Footnotes and cell styles can be precisely added through a location targeting system. The way in which 'gt' handles things for you means that you don't often have to worry about the fine details.

See also

Author

Maintainer: Richard Iannone rich@posit.co (ORCID)

Authors:

Joe Cheng joe@posit.co

Barret Schloerke barret@posit.co (ORCID)

Ellis Hughes ellis.h.hughes@gsk.com (ORCID)

Alexandra Lauer alexandralauer1@gmail.com (ORCID)

JooYoung Seo jseo1005@illinois.edu (ORCID)

Ken Brevoort ken@brevoort.com (ORCID)

Olivier Roy

Other contributors:

Posit Software, PBC [copyright holder, funder]