Get the LaTeX dependencies required for a gt table

When working with Rnw (Sweave) files or otherwise writing LaTeX code, including a gt table can be problematic if we don't have knowledge of the LaTeX dependencies. For the most part, these dependencies are the LaTeX packages that are required for rendering a gt table. gt_latex_dependencies() provides an object that can be used to provide the LaTeX in an Rnw file, allowing gt tables to work and not yield errors due to missing packages.

Usage

gt_latex_dependencies()

Value

An object of class knit_asis.

Details

Here is an example Rnw document that shows how gt_latex_dependencies() can be used in conjunction with a gt table:

%!sweave=knitr

\documentclass{article}

<<echo=FALSE>>=
library(gt)
 @

<<results='asis', echo=FALSE>>=
gt_latex_dependencies()
 @

\begin{document}

<<results='asis', echo=FALSE>>=
gt(exibble)
 @

\end{document}

Function ID

8-30

Function Introduced

v0.2.0.5 (March 31, 2020)

See also