gt_latex_dependencies {gt}	R Documentation

Get the LaTeX dependencies required for a gt table

Description

When working with Rnw (Sweave) files or otherwise writing LaTeX code, including a gt table can be problematic if we don't have knowledge of the LaTeX dependencies. For the most part, these dependencies are the LaTeX packages that are required for rendering a gt table. The gt_latex_dependencies() function provides an object that can be used to provide the LaTeX in an Rnw file, allowing gt tables to work and not yield errors due to missing packages.

Usage

gt_latex_dependencies()

Details

Here is an example Rnw document that shows how the gt_latex_dependencies() can be used in conjunction with a gt table:

%!sweave=knitr

\documentclass{article}

<<echo=FALSE>>=
library(gt)
 @

<<results='asis', echo=FALSE>>=
gt_latex_dependencies()
 @

\begin{document}

<<results='asis', echo=FALSE>>=
exibble 
 @

\end{document}

See Also

Other helper functions: cells_styles, escape_latex, html, md, pct, px

[Package gt version 0.1.0 Index]