gtcars {gt}	R Documentation

Deluxe automobiles from the 2014-2017 period

Description

Expensive and fast cars. Not your father's mtcars. Each row describes a car of a certain make, model, year, and trim. Basic specifications such as horsepower, torque, EPA MPG ratings, type of drivetrain, and transmission characteristics are provided. The country of origin for the car manufacturer is also given.

Usage

gtcars

Format

A tibble with 47 rows and 15 variables:

mfr

The name of the car manufacturer

model

The car's model name

year

The car's model year

trim

A short description of the car model's trim

bdy_style

An identifier of the car's body style, which is either coupe, convertible, sedan, or hatchback

hp, hp_rpm

The car's horsepower and the associated RPM level

trq, trq_rpm

The car's torque and the associated RPM level

mpg_c, mpg_h

The miles per gallon fuel efficiency rating for city and highway driving

drivetrain

The car's drivetrain which, for this dataset is either rwd (Rear Wheel Drive) or awd (All Wheel Drive)

trsmn

The codified transmission type, where the number part is the number of gears; the car could have automatic transmission (a), manual transmission (m), an option to switch between both types (am), or, direct drive (dd)

ctry_origin

The country name for where the vehicle manufacturer is headquartered

Source

Various pages within https://www.edmunds.com.

Examples

# Here is a glimpse at the data
# available in `gtcars`
dplyr::glimpse(gtcars)

[Package gt version 0.1.0 Index]