render_gt {gt}	R Documentation

A gt display table render function for use in Shiny

Description

A gt display table render function for use in Shiny

Usage

render_gt(expr, env = parent.frame(), quoted = FALSE,
  outputArgs = list())

Arguments

expr
an expression that creates a gt table object.

env
the environment in which to evaluate the expr.

quoted
is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

outputArgs
A list of arguments to be passed through to the implicit call to gt_output() when render_gt is used in an interactive R Markdown document.

See Also

gt_output()

Other Shiny functions: gt_output

[Package gt version 0.1.0 Index]